@@ -11,66 +11,80 @@
     "clean": "rimraf src/artifacts && rimraf src/gen && rimraf cache",
     "compile": "hardhat --max-memory 4096 compile",
     "adapter": "yarn adapter:gen && yarn adapter:build",
-    "adapter:gen": "rimraf src/gen/typechain && typechain --target ethers-v5 --outDir src/gen/typechain \"./src/artifacts/contracts/**/*[^dbg].json\"",
+    "adapter:gen": "rimraf src/gen/typechain && typechain --target ethers-v5 --out-dir src/gen/typechain \"./src/artifacts/contracts/**/*[^dbg].json\"",
     "adapter:build": "rimraf src/gen/adapter && tsc ./src/gen/typechain/index.ts --outDir ./src/gen/adapter",
-    "test": "yarn build && yarn concurrently -k --success first \"yarn start:ganache\" \"yarn test:ganache\"",
+    "test": "yarn build && yarn test:hardhat",
     "test:hardhat": "hardhat test \"tests/index.spec.ts\"",
-    "test:ganache": "hardhat --network ganache test tests/index.spec.ts",
     "test:gas-report": "REPORT_GAS=true yarn test",
-    "test:benchmark": "hardhat --network ganache test tests/index.bench.ts",
+    "test:benchmark": "hardhat test tests/index.bench.ts",
+    "test:forge": "forge test -vv",
+    "test:forge:coverage": "forge coverage --report summary",
     "lint": "yarn lint:ts && yarn lint:sol",
     "lint:fix": "yarn lint:ts:fix && yarn lint:sol:fix",
-    "lint:sol": "solhint \"./src/contracts/**/*.sol\"",
-    "lint:sol:fix": "solhint \"./src/contracts/**/*.sol\" --fix",
+    "lint:sol": "solhint \"./src/contracts/**/*.sol\" \"./tests_foundry/**/*.sol\"",
+    "lint:sol:fix": "solhint \"./src/contracts/**/*.sol\" \"./tests_foundry/**/*.sol\" --fix",
     "lint:ts": "eslint -c .eslintrc.js \"./**/*.ts\"",
     "lint:ts:fix": "eslint -c .eslintrc.js --fix \"./**/*.ts\"",
-    "format": "prettier --write ./**/*.ts",
-    "start:ganache": "cross-var ganache-cli --chainId $npm_package_config_ganacheChainID --networkId $npm_package_config_ganacheChainID --port $npm_package_config_ganachePort --gasLimit $npm_package_config_ganacheGasLimit --gasPrice $npm_package_config_ganacheGasPrice --defaultBalanceEther $npm_package_config_etherBalance --deterministic --mnemonic \"$npm_package_config_mnemonic\" $npm_package_config_extra",
-    "start:ganache:verbose": "yarn run start:ganache --verbose",
-    "stop:ganache": "ps aux | grep ganache-cli | grep -v grep | awk '{print $2}' | xargs kill -9",
+    "format": "yarn format:ts && yarn format:sol",
+    "format:ts": "prettier --write \"./**/*.ts\"",
+    "format:sol": "forge fmt",
     "release": "yarn publish src --access public",
     "release-list": "cd src && yarn release-list",
     "verify": "hardhat tenderly:verify"
   },
   "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.0.2",
-    "@nomiclabs/hardhat-truffle5": "^2.0.0",
+    "@nomiclabs/hardhat-ethers": "^2.2.2",
+    "@nomiclabs/hardhat-truffle5": "^2.0.7",
     "@nomiclabs/hardhat-web3": "^2.0.0",
-    "@tenderly/hardhat-tenderly": "^1.0.11",
-    "@typechain/ethers-v5": "^6.0.4",
-    "@types/chai-as-promised": "^7.1.0",
-    "@types/chai-string": "^1.4.1",
-    "@types/mocha": "^8.2.1",
-    "@types/node": "^14.14.35",
-    "@typescript-eslint/eslint-plugin": "^4.18.0",
-    "@typescript-eslint/parser": "^4.18.0",
-    "chai": "^4.3.4",
+    "@tenderly/hardhat-tenderly": "^1.6.1",
+    "@typechain/ethers-v5": "^10.2.0",
+    "@types/chai-as-promised": "^7.1.5",
+    "@types/chai-string": "^1.4.2",
+    "@types/mocha": "^10.0.1",
+    "@types/node": "^18.15.2",
+    "@typescript-eslint/eslint-plugin": "^5.55.0",
+    "@typescript-eslint/parser": "^5.55.0",
+    "chai": "^4.3.7",
     "chai-as-promised": "^7.1.1",
-    "chai-bignumber": "^3.0.0",
+    "chai-bignumber": "^3.1.0",
     "chai-string": "^1.5.0",
-    "concurrently": "^6.0.0",
+    "concurrently": "^7.6.0",
     "cross-var": "^1.1.0",
-    "dotenv": "^8.2.0",
-    "eslint": "^7.22.0",
-    "eslint-config-prettier": "^8.1.0",
-    "eslint-plugin-import": "^2.22.0",
-    "eslint-plugin-prettier": "^3.3.1",
-    "ethers": "^5.0.32",
-    "ganache-cli": "6.12.2",
-    "hardhat": "2.1.1",
-    "hardhat-gas-reporter": "1.0.4",
-    "rimraf": "^3.0.2",
-    "solhint": "^3.3.4",
-    "ts-node": "^9.1.1",
-    "typechain": "^4.0.3",
-    "typescript": "4.2.3"
+    "dotenv": "^16.0.3",
+    "eslint": "^8.36.0",
+    "eslint-config-prettier": "^8.7.0",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-prettier": "^4.2.1",
+    "ethers": "^5.7.2",
+    "hardhat": "2.13.0",
+    "hardhat-gas-reporter": "1.0.9",
+    "husky": "^4.2.3",
+    "lint-staged": "^13.2.0",
+    "rimraf": "^4.4.0",
+    "solhint": "^3.4.1",
+    "ts-node": "^10.9.1",
+    "typechain": "^8.1.1",
+    "typescript": "4.9.5"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "yarn lint-staged"
+    }
+  },
+  "lint-staged": {
+    "*.ts": [
+      "eslint -c .eslintrc.js",
+      "prettier --write",
+      "eslint -c .eslintrc.js"
+    ],
+    "*.sol": [
+      "yarn lint:sol",
+      "yarn format:sol",
+      "yarn lint:sol"
+    ]
   },
   "config": {
     "mnemonic": "test test test test test test test test test test test junk",
-    "ganacheChainID": 127001,
-    "ganachePort": 8545,
-    "ganacheGasLimit": "0xfffffffffff",
-    "ganacheGasPrice": "20000000000",
     "etherBalance": "100000",
     "extra": ""
   }
