@@ -1,6 +1,6 @@
 import * as chai from 'chai'
 import chaiAsPromised from 'chai-as-promised'
-const ChaiBigNumber = require('chai-bignumber')
+import ChaiBigNumber from 'chai-bignumber'
 import chaiString from 'chai-string'
 import * as ethers from 'ethers'
 
@@ -10,8 +10,4 @@ export * from './helpers'
 const BigNumber = ethers.BigNumber
 export { BigNumber }
 
-export const { assert, expect } = chai
-  .use(chaiString)
-  .use(chaiAsPromised)
-  .use(ChaiBigNumber())
-  .use(require('chai-bignumber')(BigNumber));
+export const { assert, expect } = chai.use(chaiString).use(chaiAsPromised).use(ChaiBigNumber()).use(ChaiBigNumber(BigNumber))
