@@ -1,13 +1,13 @@
 import { HardhatUserConfig } from 'hardhat/config'
 
 import '@nomiclabs/hardhat-truffle5'
-import "@nomiclabs/hardhat-ethers"
+import '@nomiclabs/hardhat-ethers'
 import 'hardhat-gas-reporter'
 import '@tenderly/hardhat-tenderly'
 
 const config: HardhatUserConfig = {
   solidity: {
-    version: '0.7.4',
+    version: '0.8.16', // Matches forge version
     settings: {
       optimizer: {
         enabled: true,
@@ -23,16 +23,12 @@ const config: HardhatUserConfig = {
     tests: '../tests'
   },
   networks: {
-    ganache: {
-      url: 'http://127.0.0.1:8545',
-      blockGasLimit: 10000000
-    },
     matic: {
       url: 'https://rpc-mainnet.matic.network'
     },
     hardhat: {
       blockGasLimit: 0xfffffffffff,
-      gasPrice: 20000000000,
+      gasPrice: 20000000000
     }
   },
   gasReporter: {
