@@ -1,43 +1,43 @@
-import * as ethers from 'ethers'
+import { ethers, BigNumber } from 'ethers'
 
 import {
   AbstractContract,
+  ArrayAccessError,
+  CallError,
+  createTestWallet,
   expect,
-  OpCodeError,
-  RevertError,
-  getSellTokenData20,
   getAddLiquidityData,
   getRemoveLiquidityData,
-  getBuyTokenData,
+  getSellTokenData20,
+  RevertError,
+  RevertUnsafeMathError
 } from './utils'
 
-import * as utils from './utils'
-
 import {
-  ERC20TokenMock,
   ERC1155Mock,
-  ERC1155RoyaltyMock,
   ERC1155PackedBalanceMock,
+  ERC1155RoyaltyMock,
+  ERC20TokenMock,
   NiftyswapExchange20,
   NiftyswapFactory20
 } from 'src/gen/typechain'
 
 import { abi as exchangeABI } from '@0xsequence/niftyswap/artifacts/contracts/exchange/NiftyswapExchange20.sol/NiftyswapExchange20.json'
-import { BigNumber } from 'ethers'
+import { Interface } from 'ethers/lib/utils'
 import { web3 } from 'hardhat'
 import { ERC1155MetadataPrefix } from 'src/gen/typechain/ERC1155MetadataPrefix'
+import { EventFilter } from './typings/ethers'
 
-const exchangeIface = new ethers.utils.Interface(exchangeABI)
+const { One } = ethers.constants
+
+const exchangeIface = new Interface(exchangeABI)
 
 // init test wallets from package.json mnemonic
 
-const { wallet: ownerWallet, provider: ownerProvider, signer: ownerSigner } = utils.createTestWallet(web3, 0)
-
-const { wallet: userWallet, provider: userProvider, signer: userSigner } = utils.createTestWallet(web3, 2)
-
-const { wallet: operatorWallet, provider: operatorProvider, signer: operatorSigner } = utils.createTestWallet(web3, 4)
-
-const { wallet: randomWallet, provider: randomProvider, signer: randomSigner } = utils.createTestWallet(web3, 5)
+const { wallet: ownerWallet } = createTestWallet(web3, 0)
+const { wallet: userWallet, signer: userSigner } = createTestWallet(web3, 2)
+const { wallet: operatorWallet, provider: operatorProvider, signer: operatorSigner } = createTestWallet(web3, 4)
+const { wallet: randomWallet } = createTestWallet(web3, 5)
 
 const getBig = (id: number) => BigNumber.from(id)
 
@@ -79,26 +79,24 @@ describe('NiftyswapExchange20', () => {
   const currencyAmount = BigNumber.from(10000000).mul(BigNumber.from(10).pow(18))
 
   // Fees param
-  let LP_FEE
-  let LP_FEE_MULTIPLIER
-  const ROYALTY_FEE = 200               // 2%
-  const EXTRA_FEE = 66667777            // flat fee
+  let LP_FEE: number
+  let LP_FEE_MULTIPLIER: number
+  const ROYALTY_FEE = 200 // 2%
+  const EXTRA_FEE = 66667777 // flat fee
 
   // Add liquidity data
   const tokenAmountToAdd = BigNumber.from(300)
-  const currencyAmountToAdd = BigNumber.from(10)
-    .pow(18)
-    .mul(299)
+  const currencyAmountToAdd = BigNumber.from(10).pow(18).mul(299)
 
   // Transactions parameters
   const TX_PARAM = { gasLimit: 5000000 }
   const deadline = Math.floor(Date.now() / 1000) + 100000
 
   // Arrays
-  const types = new Array(nTokenTypes).fill('').map((a, i) => getBig(i))
-  const values = new Array(nTokenTypes).fill('').map((a, i) => nTokensPerType)
-  const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => currencyAmountToAdd)
-  const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => tokenAmountToAdd)
+  const types = new Array(nTokenTypes).fill('').map((_a, i) => getBig(i))
+  const values = new Array(nTokenTypes).fill('').map(() => nTokensPerType)
+  const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => currencyAmountToAdd)
+  const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => tokenAmountToAdd)
   const addLiquidityData: string = getAddLiquidityData(currencyAmountsToAdd, deadline)
 
   // load contract abi and deploy to test server
@@ -116,14 +114,14 @@ describe('NiftyswapExchange20', () => {
     erc1155MetadataPrefixAbstract = await AbstractContract.fromArtifactName('ERC1155MetadataPrefix')
   })
 
-  let conditions = [
+  const conditions = [
     ['Regular ERC1155 Token', 1],
     ['Packed Balance ERC1155 Token', 2],
     ['ERC-2981 Token with 0% royalty', 3],
     ['ERC-2981 Token Contract with 2% royalty', 4],
     ['Regular Token Contract with 2% royalty', 5],
     ['ERC-2981 Token Contract with 2% royalty and fixed extra fee', 6],
-    ['ERC-2981 Token Contract with 2% royalty and fixed extra fee with 10% LP fee', 7],
+    ['ERC-2981 Token Contract with 2% royalty and fixed extra fee with 10% LP fee', 7]
   ]
 
   let erc1155_error_prefix
@@ -134,7 +132,7 @@ describe('NiftyswapExchange20', () => {
   let extraFeeArray: BigNumber[] = []
   let extraFeeRecipients: string[] = []
 
-  conditions.forEach(function(condition) {
+  conditions.forEach(function (condition) {
     context(condition[0] as string, () => {
       // deploy before each test, to reset state of contract
       beforeEach(async () => {
@@ -145,21 +143,18 @@ describe('NiftyswapExchange20', () => {
           userERC1155Contract = (await ownerERC1155Contract.connect(userSigner)) as ERC1155Mock
           erc1155_error_prefix = 'ERC1155#'
           royaltyFee = BigNumber.from(0)
-
         } else if (condition[1] == 2) {
           ownerERC1155Contract = (await erc1155PackedAbstract.deploy(ownerWallet)) as ERC1155PackedBalanceMock
           operatorERC1155Contract = (await ownerERC1155Contract.connect(operatorSigner)) as ERC1155PackedBalanceMock
           userERC1155Contract = (await ownerERC1155Contract.connect(userSigner)) as ERC1155PackedBalanceMock
           erc1155_error_prefix = 'ERC1155PackedBalance#'
           royaltyFee = BigNumber.from(0)
-
         } else if (condition[1] == 3) {
           ownerERC1155Contract = (await erc1155RoyaltyAbstract.deploy(ownerWallet)) as ERC1155RoyaltyMock
           operatorERC1155Contract = (await ownerERC1155Contract.connect(operatorSigner)) as ERC1155RoyaltyMock
           userERC1155Contract = (await ownerERC1155Contract.connect(userSigner)) as ERC1155RoyaltyMock
           erc1155_error_prefix = 'ERC1155#'
           royaltyFee = BigNumber.from(0)
-          
         } else if (condition[1] == 4 || condition[1] == 6 || condition[1] == 7) {
           ownerERC1155Contract = (await erc1155RoyaltyAbstract.deploy(ownerWallet)) as ERC1155RoyaltyMock
           operatorERC1155Contract = (await ownerERC1155Contract.connect(operatorSigner)) as ERC1155RoyaltyMock
@@ -171,7 +166,7 @@ describe('NiftyswapExchange20', () => {
           await (ownerERC1155Contract as ERC1155RoyaltyMock).functions.setFeeRecipient(randomAddress)
 
           erc1155_error_prefix = 'ERC1155#'
-        } 
+        }
 
         // Deploy Currency Token contract
         ownerCurrencyContract = (await erc20Abstract.deploy(ownerWallet)) as ERC20TokenMock
@@ -186,7 +181,7 @@ describe('NiftyswapExchange20', () => {
         } else {
           LP_FEE = 10 // 1%
         }
-        LP_FEE_MULTIPLIER = 1000-LP_FEE // 1%
+        LP_FEE_MULTIPLIER = 1000 - LP_FEE
 
         // Create exchange contract for the ERC-20/1155 token
         await niftyswapFactoryContract.functions.createExchange(
@@ -210,11 +205,11 @@ describe('NiftyswapExchange20', () => {
         niftyswapExchangeContract = new ethers.Contract(exchangeAddress, exchangeABI, ownerWallet) as NiftyswapExchange20
         operatorExchangeContract = (await niftyswapExchangeContract.connect(operatorSigner)) as NiftyswapExchange20
         userExchangeContract = (await niftyswapExchangeContract.connect(userSigner)) as NiftyswapExchange20
-        
+
         // Set royalty for condition 5
         if (condition[1] == 5) {
           royaltyFee = BigNumber.from(ROYALTY_FEE)
-          await niftyswapExchangeContract.functions.setRoyaltyInfo(ROYALTY_FEE, randomAddress, {gasLimit: 8000000})
+          await niftyswapExchangeContract.functions.setRoyaltyInfo(ROYALTY_FEE, randomAddress, { gasLimit: 8000000 })
         }
 
         // Set extra fees for condition 6
@@ -292,35 +287,35 @@ describe('NiftyswapExchange20', () => {
 
         describe('getBuyPrice() function', () => {
           it('should round UP', async () => {
-            let bought_amount = 100
-            let sellReserve = 1500
-            let buyReserve = 751
+            const bought_amount = 100
+            const sellReserve = 1500
+            const buyReserve = 751
 
             // Calculate the price manually with floats
             const numerator = bought_amount * sellReserve * 1000
-            const denominator =  (buyReserve - bought_amount) * LP_FEE_MULTIPLIER
+            const denominator = (buyReserve - bought_amount) * LP_FEE_MULTIPLIER
             const floatPrice = numerator / denominator // Price without rounding
 
             // Get the price from the contract
             const price = await niftyswapExchangeContract.functions.getBuyPrice(bought_amount, sellReserve, buyReserve)
-            expect(price[0]).to.be.eql(BigNumber.from(Math.ceil(floatPrice))) 
+            expect(price[0]).to.be.eql(BigNumber.from(Math.ceil(floatPrice)))
           })
         })
 
         describe('getSellPrice() function', () => {
           it('should round DOWN', async () => {
-            let sold_amount = 100
-            let buyReserve = 1500
-            let sellReserve = 751 
+            const sold_amount = 100
+            const buyReserve = 1500
+            const sellReserve = 751
 
             // Calculate price manually with floats
             const numerator = sold_amount * LP_FEE_MULTIPLIER * buyReserve
-            const denominator =  sellReserve * 1000 + sold_amount * LP_FEE_MULTIPLIER
+            const denominator = sellReserve * 1000 + sold_amount * LP_FEE_MULTIPLIER
             const floatPrice = numerator / denominator // Price without rounding
 
             // Get the price from the contract directly
             const price = await niftyswapExchangeContract.functions.getSellPrice(sold_amount, sellReserve, buyReserve)
-            expect(price[0]).to.be.eql(BigNumber.from(Math.floor(floatPrice))) 
+            expect(price[0]).to.be.eql(BigNumber.from(Math.floor(floatPrice)))
           })
         })
 
@@ -368,8 +363,8 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should ROUND UP the currency amount to be deposited on second deposit', async () => {
-          let addLiquidityData1 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
-          let tokenAmountsToAdd1 = [BigNumber.from(2)]
+          const addLiquidityData1 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
+          const tokenAmountsToAdd1 = [BigNumber.from(2)]
 
           // After 2nd deposit
           await operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -381,11 +376,11 @@ describe('NiftyswapExchange20', () => {
             { gasLimit: 50000000 }
           )
 
-          let reserve1 = (await operatorExchangeContract.functions.getCurrencyReserves([0]))[0]
+          const reserve1 = (await operatorExchangeContract.functions.getCurrencyReserves([0]))[0]
           expect(reserve1[0]).to.be.eql(BigNumber.from(1000000001))
 
-          let addLiquidityData2 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
-          let tokenAmountsToAdd2 = [BigNumber.from(1)] // 1 less
+          const addLiquidityData2 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
+          const tokenAmountsToAdd2 = [One] // 1 less
 
           // After 2nd deposit
           await operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -397,13 +392,13 @@ describe('NiftyswapExchange20', () => {
             { gasLimit: 50000000 }
           )
 
-          let reserve2 = (await operatorExchangeContract.functions.getCurrencyReserves([0]))[0]
+          const reserve2 = (await operatorExchangeContract.functions.getCurrencyReserves([0]))[0]
           expect(reserve2[0]).to.be.eql(BigNumber.from(1500000002)) // Should be 1500000001.5
         })
 
         it('should ROUND DOWN the amount of liquidity to mint on second deposit', async () => {
-          let addLiquidityData1 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
-          let tokenAmountsToAdd1 = [BigNumber.from(2)]
+          const addLiquidityData1 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
+          const tokenAmountsToAdd1 = [BigNumber.from(2)]
 
           // first deposit
           await operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -415,11 +410,11 @@ describe('NiftyswapExchange20', () => {
             { gasLimit: 50000000 }
           )
 
-          let liquidity_supply1 = (await operatorExchangeContract.functions.getTotalSupply([0]))[0]
+          const liquidity_supply1 = (await operatorExchangeContract.functions.getTotalSupply([0]))[0]
           expect(liquidity_supply1[0]).to.be.eql(BigNumber.from(1000000001))
 
-          let addLiquidityData2 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
-          let tokenAmountsToAdd2 = [BigNumber.from(1)] // 1 less
+          const addLiquidityData2 = getAddLiquidityData([BigNumber.from(1000000001)], deadline)
+          const tokenAmountsToAdd2 = [One] // 1 less
 
           // After 2nd deposit
           await operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -431,13 +426,13 @@ describe('NiftyswapExchange20', () => {
             { gasLimit: 50000000 }
           )
 
-          let liquidity_supply2 = (await operatorExchangeContract.functions.getTotalSupply([0]))[0]
+          const liquidity_supply2 = (await operatorExchangeContract.functions.getTotalSupply([0]))[0]
           expect(liquidity_supply2[0]).to.be.eql(BigNumber.from(1500000001)) // Should be 1500000001.5
         })
 
         it('should REVERT if deadline is passed', async () => {
-          let timestamp = Math.floor(Date.now() / 1000) - 1
-          let addLiquidityData = getAddLiquidityData(currencyAmountsToAdd, timestamp)
+          const timestamp = Math.floor(Date.now() / 1000) - 1
+          const addLiquidityData = getAddLiquidityData(currencyAmountsToAdd, timestamp)
 
           const tx = operatorERC1155Contract.functions.safeBatchTransferFrom(
             operatorAddress,
@@ -451,9 +446,9 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should REVERT if a maxCurrency is null', async () => {
-          let currencyAmountsToAdd = new Array(nTokenTypes).fill('').map((a, i) => currencyAmountToAdd)
+          const currencyAmountsToAdd = new Array(nTokenTypes).fill('').map(() => currencyAmountToAdd)
           currencyAmountsToAdd[5] = BigNumber.from(0)
-          let addLiquidityData = getAddLiquidityData(currencyAmountsToAdd, deadline)
+          const addLiquidityData = getAddLiquidityData(currencyAmountsToAdd, deadline)
 
           const tx = operatorERC1155Contract.functions.safeBatchTransferFrom(
             operatorAddress,
@@ -467,7 +462,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should REVERT if a token amount is null', async () => {
-          let tokenAmountsToAddCopy = [...tokenAmountsToAdd]
+          const tokenAmountsToAddCopy = [...tokenAmountsToAdd]
           tokenAmountsToAddCopy[5] = BigNumber.from(0)
 
           const tx = operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -482,7 +477,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should REVERT if arrays are not the same length', async () => {
-          let currencyAmount1 = currencyAmountToAdd.add(1)
+          const currencyAmount1 = currencyAmountToAdd.add(1)
 
           // If expected tier is larger, then should be fine
           let data = getAddLiquidityData([currencyAmountToAdd, currencyAmountToAdd, currencyAmountToAdd], deadline)
@@ -506,7 +501,7 @@ describe('NiftyswapExchange20', () => {
             data,
             TX_PARAM
           )
-          await expect(tx2).to.be.rejectedWith(OpCodeError())
+          await expect(tx2).to.be.rejectedWith(ArrayAccessError())
 
           data = getAddLiquidityData([currencyAmount1, currencyAmount1], deadline)
           const tx3 = operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -556,7 +551,7 @@ describe('NiftyswapExchange20', () => {
             data,
             TX_PARAM
           )
-          await expect(tx6).to.be.rejectedWith(OpCodeError())
+          await expect(tx6).to.be.rejectedWith(ArrayAccessError())
         })
 
         it('should REVERT if any duplicate', async () => {
@@ -568,9 +563,7 @@ describe('NiftyswapExchange20', () => {
             addLiquidityData,
             { gasLimit: 50000000 }
           )
-          await expect(tx1).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx1).to.be.rejectedWith(RevertError('NE20#32'))
 
           const tx2 = operatorERC1155Contract.functions.safeBatchTransferFrom(
             operatorAddress,
@@ -580,9 +573,7 @@ describe('NiftyswapExchange20', () => {
             addLiquidityData,
             { gasLimit: 50000000 }
           )
-          await expect(tx2).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx2).to.be.rejectedWith(RevertError('NE20#32'))
 
           const tx3 = operatorERC1155Contract.functions.safeBatchTransferFrom(
             operatorAddress,
@@ -592,13 +583,11 @@ describe('NiftyswapExchange20', () => {
             addLiquidityData,
             { gasLimit: 50000000 }
           )
-          await expect(tx3).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx3).to.be.rejectedWith(RevertError('NE20#32'))
         })
 
         context('When liquidity was added', () => {
-          const currencyAmountsToAddOne = new Array(nTokenTypes).fill('').map((a, i) => currencyAmountToAdd.add(1))
+          const currencyAmountsToAddOne = new Array(nTokenTypes).fill('').map(() => currencyAmountToAdd.add(1))
           const addLiquidityDataOne = getAddLiquidityData(currencyAmountsToAddOne, deadline)
 
           beforeEach(async () => {
@@ -658,7 +647,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should DECREASE the BUY prices for 2ND deposit', async () => {
-            const ones = new Array(nTokenTypes).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypes).fill('').map(() => 1)
 
             // After 1st deposit
             const prePrices = await niftyswapExchangeContract.functions.getPrice_currencyToToken(types, ones)
@@ -680,7 +669,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should DECREASE the BUY prices for 3RD deposit', async () => {
-            const ones = new Array(nTokenTypes).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypes).fill('').map(() => 1)
 
             // After 2nd deposit
             await operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -710,7 +699,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should INCREASE the SELL prices for 2ND deposit', async () => {
-            const ones = new Array(nTokenTypes).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypes).fill('').map(() => 1)
 
             // After 1st deposit
             const prePrices = await niftyswapExchangeContract.functions.getPrice_tokenToCurrency(types, ones)
@@ -732,7 +721,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should INCREASE the SELL prices for 3RD deposit', async () => {
-            const ones = new Array(nTokenTypes).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypes).fill('').map(() => 1)
 
             // After 2nd deposit
             await operatorERC1155Contract.functions.safeBatchTransferFrom(
@@ -762,15 +751,17 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should emit LiquidityAdded event', async () => {
-            let filterFromOperatorContract: ethers.ethers.EventFilter
-
             // Get event filter to get internal tx event
-            filterFromOperatorContract = niftyswapExchangeContract.filters.LiquidityAdded(null, null, null, null)
+            const filterFromOperatorContract: EventFilter = niftyswapExchangeContract.filters.LiquidityAdded(
+              null,
+              null,
+              null,
+              null
+            )
 
             // Get logs from internal transaction event
-            // @ts-ignore (https://github.com/ethers-io/ethers.js/issues/204#issuecomment-427059031)
             filterFromOperatorContract.fromBlock = 0
-            let logs = await operatorProvider.getLogs(filterFromOperatorContract)
+            const logs = await operatorProvider.getLogs(filterFromOperatorContract)
             expect(logs[0].topics[0]).to.be.eql(
               niftyswapExchangeContract.interface.getEventTopic(
                 niftyswapExchangeContract.interface.events['LiquidityAdded(address,uint256[],uint256[],uint256[])']
@@ -788,7 +779,11 @@ describe('NiftyswapExchange20', () => {
               let prefixMetadataContract: ERC1155MetadataPrefix
 
               beforeEach(async () => {
-                prefixMetadataContract = (await erc1155MetadataPrefixAbstract.deploy(ownerWallet, ["", true, ownerWallet.address])) as ERC1155MetadataPrefix
+                prefixMetadataContract = (await erc1155MetadataPrefixAbstract.deploy(ownerWallet, [
+                  '',
+                  true,
+                  ownerWallet.address
+                ])) as ERC1155MetadataPrefix
                 await niftyswapFactoryContract.setMetadataContract(prefixMetadataContract.address)
               })
 
@@ -798,7 +793,7 @@ describe('NiftyswapExchange20', () => {
               })
 
               it('should return prefixed `id@address` of token', async () => {
-                await prefixMetadataContract.functions.setUriPrefix("https://sequence.app/")
+                await prefixMetadataContract.functions.setUriPrefix('https://sequence.app/')
                 const uri = (await niftyswapExchangeContract.functions.uri(1))[0]
                 expect(uri).to.be.eql(`https://sequence.app/1@${niftyswapExchangeContract.address.toLowerCase()}`)
               })
@@ -810,9 +805,13 @@ describe('NiftyswapExchange20', () => {
               })
 
               it('should change metadata implementation', async () => {
-                const prefixMetadataContract2 = (await erc1155MetadataPrefixAbstract.deploy(ownerWallet, ["", false, ownerWallet.address])) as ERC1155MetadataPrefix
+                const prefixMetadataContract2 = (await erc1155MetadataPrefixAbstract.deploy(ownerWallet, [
+                  '',
+                  false,
+                  ownerWallet.address
+                ])) as ERC1155MetadataPrefix
                 await niftyswapFactoryContract.setMetadataContract(prefixMetadataContract2.address)
-                await prefixMetadataContract2.functions.setUriPrefix("https://v2.sequence.app/")
+                await prefixMetadataContract2.functions.setUriPrefix('https://v2.sequence.app/')
                 const uri = (await niftyswapExchangeContract.functions.uri(25666))[0]
                 expect(uri).to.be.eql(`https://v2.sequence.app/25666`)
               })
@@ -821,7 +820,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         context('When liquidity was added for the second time', () => {
-          const currencyAmountsToAddOne = new Array(nTokenTypes).fill('').map((a, i) => currencyAmountToAdd.add(1))
+          const currencyAmountsToAddOne = new Array(nTokenTypes).fill('').map(() => currencyAmountToAdd.add(1))
           const addLiquidityDataOne = getAddLiquidityData(currencyAmountsToAddOne, deadline)
 
           beforeEach(async () => {
@@ -844,9 +843,9 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should REVERT if a maxCurrency is exceeded', async () => {
-            let currencyAmountsToAdd = new Array(nTokenTypes).fill('').map((a, i) => currencyAmountToAdd)
+            const currencyAmountsToAdd = new Array(nTokenTypes).fill('').map(() => currencyAmountToAdd)
             currencyAmountsToAdd[5] = BigNumber.from(1000)
-            let addLiquidityData = getAddLiquidityData(currencyAmountsToAdd, deadline)
+            const addLiquidityData = getAddLiquidityData(currencyAmountsToAdd, deadline)
 
             const tx = operatorERC1155Contract.functions.safeBatchTransferFrom(
               operatorAddress,
@@ -934,12 +933,12 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should pass when balances are sufficient', async () => {
-            let maxCurrency: BigNumber[] = []
+            const maxCurrency: BigNumber[] = []
 
             for (let i = 0; i < nTokenTypes; i++) {
               maxCurrency.push(currencyAmountToAdd.mul(2))
             }
-            let addLiquidityData2 = getAddLiquidityData(maxCurrency, deadline)
+            const addLiquidityData2 = getAddLiquidityData(maxCurrency, deadline)
 
             const tx = operatorERC1155Contract.functions.safeBatchTransferFrom(
               operatorAddress,
@@ -958,17 +957,14 @@ describe('NiftyswapExchange20', () => {
         const nTokenTypesToRemove = 30
 
         const tokenAmountToRemove = BigNumber.from(75)
-        const currencyAmountToRemove = BigNumber.from(10)
-          .pow(18)
-          .mul(299)
-          .div(4)
+        const currencyAmountToRemove = BigNumber.from(10).pow(18).mul(299).div(4)
 
-        const typesToRemove = new Array(nTokenTypesToRemove).fill('').map((a, i) => getBig(i))
+        const typesToRemove = new Array(nTokenTypesToRemove).fill('').map((_a, i) => getBig(i))
 
-        const tokenAmountsToRemove = new Array(nTokenTypesToRemove).fill('').map((a, i) => tokenAmountToRemove)
-        const currencyAmountsToRemove = new Array(nTokenTypesToRemove).fill('').map((a, i) => currencyAmountToRemove)
+        const tokenAmountsToRemove = new Array(nTokenTypesToRemove).fill('').map(() => tokenAmountToRemove)
+        const currencyAmountsToRemove = new Array(nTokenTypesToRemove).fill('').map(() => currencyAmountToRemove)
 
-        const niftyTokenToSend = new Array(nTokenTypesToRemove).fill('').map((a, i) => currencyAmountToRemove)
+        const niftyTokenToSend = new Array(nTokenTypesToRemove).fill('').map(() => currencyAmountToRemove)
 
         const removeLiquidityData: string = getRemoveLiquidityData(currencyAmountsToRemove, tokenAmountsToRemove, deadline)
 
@@ -981,11 +977,11 @@ describe('NiftyswapExchange20', () => {
             removeLiquidityData,
             { gasLimit: 8000000 }
           )
-          await expect(tx).to.be.rejectedWith(RevertError('SafeMath#sub: UNDERFLOW'))
+          await expect(tx).to.be.rejectedWith(RevertUnsafeMathError())
         })
 
         it('should revert if empty reserve', async () => {
-          const zeroArray = new Array(nTokenTypesToRemove).fill('').map((a, i) => BigNumber.from(0))
+          const zeroArray = new Array(nTokenTypesToRemove).fill('').map(() => BigNumber.from(0))
           const tx = operatorExchangeContract.functions.safeBatchTransferFrom(
             operatorAddress,
             niftyswapExchangeContract.address,
@@ -999,10 +995,10 @@ describe('NiftyswapExchange20', () => {
 
         describe('Trade token rounding error', () => {
           it('Should trade rounding error when withdrawing liquidity', async () => {
-            const types = [BigNumber.from(1)]
+            const types = [One]
             const currencyAmountToAdd = [BigNumber.from(25000)]
             const tokenAmountToAdd = [BigNumber.from(100)]
-  
+
             await operatorERC1155Contract.functions.safeBatchTransferFrom(
               operatorAddress,
               niftyswapExchangeContract.address,
@@ -1013,14 +1009,16 @@ describe('NiftyswapExchange20', () => {
             )
 
             // Add a single unit of liquidity
-            await operatorERC1155Contract.connect(userWallet).functions.safeBatchTransferFrom(
-              userAddress,
-              niftyswapExchangeContract.address,
-              types,
-              [1],
-              getAddLiquidityData([BigNumber.from(250)], deadline),
-              { gasLimit: 50000000 }
-            )
+            await operatorERC1155Contract
+              .connect(userWallet)
+              .functions.safeBatchTransferFrom(
+                userAddress,
+                niftyswapExchangeContract.address,
+                types,
+                [1],
+                getAddLiquidityData([BigNumber.from(250)], deadline),
+                { gasLimit: 50000000 }
+              )
 
             let userTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
             let userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
@@ -1028,29 +1026,24 @@ describe('NiftyswapExchange20', () => {
             const maxCurrency = extraFeeArray.length === 0 ? BigNumber.from(300) : BigNumber.from(300).add(extraFeeArray[0])
 
             // Buy a single unit of token, force amount to be rounded
-            await userExchangeContract.functions.buyTokens(
-              types,
-              [BigNumber.from(1)],
-              maxCurrency,
-              deadline,
-              userAddress,
-              [],
-              [],
-              { gasLimit: 8000000 }
-            )
+            await userExchangeContract.functions.buyTokens(types, [One], maxCurrency, deadline, userAddress, [], [], {
+              gasLimit: 8000000
+            })
 
             userTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
             userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
 
             // Withdraw liquidity, it should return 0 tokens but more than 250 currency
-            await niftyswapExchangeContract.connect(userWallet).functions.safeBatchTransferFrom(
-              userAddress,
-              niftyswapExchangeContract.address,
-              types,
-              [250],
-              getRemoveLiquidityData([BigNumber.from(250)], [BigNumber.from(0)], deadline),
-              { gasLimit: 8000000 }
-            )
+            await niftyswapExchangeContract
+              .connect(userWallet)
+              .functions.safeBatchTransferFrom(
+                userAddress,
+                niftyswapExchangeContract.address,
+                types,
+                [250],
+                getRemoveLiquidityData([BigNumber.from(250)], [BigNumber.from(0)], deadline),
+                { gasLimit: 8000000 }
+              )
 
             const newUserCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
             const newUserTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
@@ -1074,10 +1067,10 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('Should trade rounding error with high slippage when withdrawing liquidity', async () => {
-            const types = [BigNumber.from(1)]
+            const types = [One]
             const currencyAmountToAdd = [BigNumber.from(1250)]
             const tokenAmountToAdd = [BigNumber.from(5)]
-  
+
             await operatorERC1155Contract.functions.safeBatchTransferFrom(
               operatorAddress,
               niftyswapExchangeContract.address,
@@ -1088,14 +1081,16 @@ describe('NiftyswapExchange20', () => {
             )
 
             // Add a single unit of liquidity
-            await operatorERC1155Contract.connect(userWallet).functions.safeBatchTransferFrom(
-              userAddress,
-              niftyswapExchangeContract.address,
-              types,
-              [1],
-              getAddLiquidityData([BigNumber.from(250)], deadline),
-              { gasLimit: 50000000 }
-            )
+            await operatorERC1155Contract
+              .connect(userWallet)
+              .functions.safeBatchTransferFrom(
+                userAddress,
+                niftyswapExchangeContract.address,
+                types,
+                [1],
+                getAddLiquidityData([BigNumber.from(250)], deadline),
+                { gasLimit: 50000000 }
+              )
 
             let userTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
             let userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
@@ -1103,29 +1098,24 @@ describe('NiftyswapExchange20', () => {
             const maxCurrency = extraFeeArray.length === 0 ? BigNumber.from(1000) : BigNumber.from(1000).add(extraFeeArray[0])
 
             // Buy a single unit of token, force amount to be rounded
-            await userExchangeContract.functions.buyTokens(
-              types,
-              [BigNumber.from(1)],
-              maxCurrency,
-              deadline,
-              userAddress,
-              [],
-              [],
-              { gasLimit: 8000000 }
-            )
+            await userExchangeContract.functions.buyTokens(types, [One], maxCurrency, deadline, userAddress, [], [], {
+              gasLimit: 8000000
+            })
 
             userTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
             userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
 
             // // Withdraw liquidity, it should return 0 tokens but more than 250 currency
-            await niftyswapExchangeContract.connect(userWallet).functions.safeBatchTransferFrom(
-              userAddress,
-              niftyswapExchangeContract.address,
-              types,
-              [250],
-              getRemoveLiquidityData([BigNumber.from(250)], [BigNumber.from(0)], deadline),
-              { gasLimit: 8000000 }
-            )
+            await niftyswapExchangeContract
+              .connect(userWallet)
+              .functions.safeBatchTransferFrom(
+                userAddress,
+                niftyswapExchangeContract.address,
+                types,
+                [250],
+                getRemoveLiquidityData([BigNumber.from(250)], [BigNumber.from(0)], deadline),
+                { gasLimit: 8000000 }
+              )
 
             const newUserCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
             const newUserTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
@@ -1149,10 +1139,10 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('Should trade rounding error when withdrawing liquidity, without rounding to zero', async () => {
-            const types = [BigNumber.from(1)]
+            const types = [One]
             const currencyAmountToAdd = [BigNumber.from(25000)]
             const tokenAmountToAdd = [BigNumber.from(100)]
-  
+
             await operatorERC1155Contract.functions.safeBatchTransferFrom(
               operatorAddress,
               niftyswapExchangeContract.address,
@@ -1163,14 +1153,16 @@ describe('NiftyswapExchange20', () => {
             )
 
             // Add a single unit of liquidity
-            await operatorERC1155Contract.connect(userWallet).functions.safeBatchTransferFrom(
-              userAddress,
-              niftyswapExchangeContract.address,
-              types,
-              [2],
-              getAddLiquidityData([BigNumber.from(500)], deadline),
-              { gasLimit: 50000000 }
-            )
+            await operatorERC1155Contract
+              .connect(userWallet)
+              .functions.safeBatchTransferFrom(
+                userAddress,
+                niftyswapExchangeContract.address,
+                types,
+                [2],
+                getAddLiquidityData([BigNumber.from(500)], deadline),
+                { gasLimit: 50000000 }
+              )
 
             let userTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
             let userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
@@ -1178,30 +1170,25 @@ describe('NiftyswapExchange20', () => {
             const maxCurrency = extraFeeArray.length === 0 ? BigNumber.from(300) : BigNumber.from(300).add(extraFeeArray[0])
 
             // Buy a single unit of token, force amount to be rounded
-            await userExchangeContract.functions.buyTokens(
-              types,
-              [BigNumber.from(1)],
-              maxCurrency,
-              deadline,
-              userAddress,
-              [],
-              [],
-              { gasLimit: 8000000 }
-            )
+            await userExchangeContract.functions.buyTokens(types, [One], maxCurrency, deadline, userAddress, [], [], {
+              gasLimit: 8000000
+            })
 
             // Get liquidity of tokens, currency and totalSupply
             userTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
             userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
 
             // Withdraw liquidity, it should return 0 tokens but more than 250 currency
-            await niftyswapExchangeContract.connect(userWallet).functions.safeBatchTransferFrom(
-              userAddress,
-              niftyswapExchangeContract.address,
-              types,
-              [500],
-              getRemoveLiquidityData([BigNumber.from(250)], [BigNumber.from(0)], deadline),
-              { gasLimit: 8000000 }
-            )
+            await niftyswapExchangeContract
+              .connect(userWallet)
+              .functions.safeBatchTransferFrom(
+                userAddress,
+                niftyswapExchangeContract.address,
+                types,
+                [500],
+                getRemoveLiquidityData([BigNumber.from(250)], [BigNumber.from(0)], deadline),
+                { gasLimit: 8000000 }
+              )
 
             const newUserCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
             const newUserTokenBalance = await userERC1155Contract.functions.balanceOf(userAddress, types[0])
@@ -1209,7 +1196,7 @@ describe('NiftyswapExchange20', () => {
             const diffUserCurrency = newUserCurrencyBalance[0].sub(userCurrencyBalance[0])
             const diffUserToken = newUserTokenBalance[0].sub(userTokenBalance[0])
 
-            expect(diffUserToken).to.be.eql(BigNumber.from(1))
+            expect(diffUserToken).to.be.eql(One)
 
             if (royaltyFee.isZero()) {
               expect(diffUserCurrency).to.be.eql(BigNumber.from(747))
@@ -1236,9 +1223,9 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should revert if insufficient currency', async () => {
-            let currencyAmountsToRemoveCopy = [...currencyAmountsToRemove]
+            const currencyAmountsToRemoveCopy = [...currencyAmountsToRemove]
             currencyAmountsToRemoveCopy[5] = BigNumber.from(currencyAmountsToRemoveCopy[5].mul(10000))
-            let removeLiquidityData = getRemoveLiquidityData(currencyAmountsToRemoveCopy, tokenAmountsToRemove, deadline)
+            const removeLiquidityData = getRemoveLiquidityData(currencyAmountsToRemoveCopy, tokenAmountsToRemove, deadline)
 
             const tx = operatorExchangeContract.functions.safeBatchTransferFrom(
               operatorAddress,
@@ -1252,9 +1239,9 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should revert if insufficient tokens', async () => {
-            let tokenAmountsToRemoveCopy = [...tokenAmountsToRemove]
+            const tokenAmountsToRemoveCopy = [...tokenAmountsToRemove]
             tokenAmountsToRemoveCopy[5] = BigNumber.from(tokenAmountsToRemoveCopy[5].mul(10000))
-            let removeLiquidityData = getRemoveLiquidityData(currencyAmountsToRemove, tokenAmountsToRemoveCopy, deadline)
+            const removeLiquidityData = getRemoveLiquidityData(currencyAmountsToRemove, tokenAmountsToRemoveCopy, deadline)
 
             const tx = operatorExchangeContract.functions.safeBatchTransferFrom(
               operatorAddress,
@@ -1276,9 +1263,7 @@ describe('NiftyswapExchange20', () => {
               removeLiquidityData,
               { gasLimit: 8000000 }
             )
-            await expect(tx1).to.be.rejectedWith(
-              RevertError('NE20#32')
-            )
+            await expect(tx1).to.be.rejectedWith(RevertError('NE20#32'))
 
             const tx2 = operatorExchangeContract.functions.safeBatchTransferFrom(
               operatorAddress,
@@ -1288,9 +1273,7 @@ describe('NiftyswapExchange20', () => {
               removeLiquidityData,
               { gasLimit: 8000000 }
             )
-            await expect(tx2).to.be.rejectedWith(
-              RevertError('NE20#32')
-            )
+            await expect(tx2).to.be.rejectedWith(RevertError('NE20#32'))
 
             const tx3 = operatorExchangeContract.functions.safeBatchTransferFrom(
               operatorAddress,
@@ -1300,9 +1283,7 @@ describe('NiftyswapExchange20', () => {
               removeLiquidityData,
               { gasLimit: 8000000 }
             )
-            await expect(tx3).to.be.rejectedWith(
-              RevertError('NE20#32')
-            )
+            await expect(tx3).to.be.rejectedWith(RevertError('NE20#32'))
           })
 
           it('should REVERT if arrays are not the same length', async () => {
@@ -1361,7 +1342,7 @@ describe('NiftyswapExchange20', () => {
               data,
               TX_PARAM
             )
-            await expect(tx5).to.be.rejectedWith(OpCodeError())
+            await expect(tx5).to.be.rejectedWith(ArrayAccessError())
 
             data = getRemoveLiquidityData([currencyAmountToRemove, currencyAmountToRemove], [tokenAmountToRemove], deadline)
             const tx6 = operatorExchangeContract.functions.safeBatchTransferFrom(
@@ -1431,7 +1412,7 @@ describe('NiftyswapExchange20', () => {
               data,
               TX_PARAM
             )
-            await expect(tx11).to.be.rejectedWith(OpCodeError())
+            await expect(tx11).to.be.rejectedWith(ArrayAccessError())
 
             data = getRemoveLiquidityData([currencyAmountToRemove], [tokenAmountToRemove, tokenAmountToRemove], deadline)
             const tx12 = operatorExchangeContract.functions.safeBatchTransferFrom(
@@ -1442,7 +1423,7 @@ describe('NiftyswapExchange20', () => {
               data,
               TX_PARAM
             )
-            await expect(tx12).to.be.rejectedWith(OpCodeError())
+            await expect(tx12).to.be.rejectedWith(ArrayAccessError())
 
             data = getRemoveLiquidityData(
               [currencyAmountToRemove, currencyAmountToRemove],
@@ -1488,7 +1469,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should INCREASE the BUY prices for 2ND withdraw', async () => {
-            const ones = new Array(nTokenTypesToRemove).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypesToRemove).fill('').map(() => 1)
 
             // After 1st deposit
             const prePrices = await niftyswapExchangeContract.functions.getPrice_currencyToToken(types, ones)
@@ -1510,7 +1491,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should INCREASE the BUY prices for 3RD withdraw', async () => {
-            const ones = new Array(nTokenTypesToRemove).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypesToRemove).fill('').map(() => 1)
 
             // After 2nd deposit
             await operatorExchangeContract.functions.safeBatchTransferFrom(
@@ -1540,7 +1521,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should DECREASE the SELL prices for 2ND withdraw', async () => {
-            const ones = new Array(nTokenTypesToRemove).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypesToRemove).fill('').map(() => 1)
 
             // After 1st deposit
             const prePrices = await niftyswapExchangeContract.functions.getPrice_tokenToCurrency(types, ones)
@@ -1562,7 +1543,7 @@ describe('NiftyswapExchange20', () => {
           })
 
           it('should DECREASE the SELL prices for 3RD withdraw', async () => {
-            const ones = new Array(nTokenTypesToRemove).fill('').map((a, i) => 1)
+            const ones = new Array(nTokenTypesToRemove).fill('').map(() => 1)
 
             // After 2nd deposit
             await operatorExchangeContract.functions.safeBatchTransferFrom(
@@ -1617,9 +1598,7 @@ describe('NiftyswapExchange20', () => {
 
             it('should update currency balances', async () => {
               const expectedVal = currencyAmountToAdd.mul(nTokenTypes).sub(currencyAmountToRemove.mul(nTokenTypes))
-              const exchangeBalance = await userCurrencyContract.functions.balanceOf(
-                niftyswapExchangeContract.address,
-              )
+              const exchangeBalance = await userCurrencyContract.functions.balanceOf(niftyswapExchangeContract.address)
               const operatorBalance = await userCurrencyContract.functions.balanceOf(operatorAddress)
 
               expect(exchangeBalance[0]).to.be.eql(expectedVal)
@@ -1658,8 +1637,8 @@ describe('NiftyswapExchange20', () => {
 
             it('should emit LiquidityRemoved event', async () => {
               const receipt = await tx.wait(1)
-              const ev = receipt.events!.pop()!
-              expect(ev.event).to.be.eql('LiquidityRemoved')
+              const ev = receipt.events?.pop()
+              expect(ev?.event).to.be.eql('LiquidityRemoved')
             })
           })
         })
@@ -1668,7 +1647,7 @@ describe('NiftyswapExchange20', () => {
       describe('_tokenToCurrency() function', () => {
         //Sell
         const tokenAmountToSell = BigNumber.from(50)
-        const tokensAmountsToSell: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => tokenAmountToSell)
+        const tokensAmountsToSell: BigNumber[] = new Array(nTokenTypes).fill('').map(() => tokenAmountToSell)
         let sellTokenData: string
         let cost: BigNumber
         let expectedRoyalty: BigNumber
@@ -1685,14 +1664,16 @@ describe('NiftyswapExchange20', () => {
             { gasLimit: 30000000 }
           )
 
-          const allcosts = (await niftyswapExchangeContract.functions.getPrice_tokenToCurrency(types, tokensAmountsToSell))
+          const allcosts = await niftyswapExchangeContract.functions.getPrice_tokenToCurrency(types, tokensAmountsToSell)
           cost = allcosts[0].reduce((acc: BigNumber, a: BigNumber) => acc.add(a)).sub(extraFee)
 
-          // Expected royalty 
-          const tokenReserve = (await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, 0))
+          // Expected royalty
+          const tokenReserve = await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, 0)
           const currencyReserve = (await niftyswapExchangeContract.functions.getCurrencyReserves([0]))[0][0]
-          preRoyaltyCost = (await niftyswapExchangeContract.functions.getSellPrice(tokenAmountToSell, tokenReserve, currencyReserve))[0]
-          expectedRoyalty = (preRoyaltyCost.mul(royaltyFee)).div(10000).mul(nTokenTypes)
+          preRoyaltyCost = (
+            await niftyswapExchangeContract.functions.getSellPrice(tokenAmountToSell, tokenReserve, currencyReserve)
+          )[0]
+          expectedRoyalty = preRoyaltyCost.mul(royaltyFee).div(10000).mul(nTokenTypes)
 
           // Sell
           sellTokenData = getSellTokenData20(userAddress, cost, deadline, extraFeeRecipients, extraFeeArray)
@@ -1709,14 +1690,14 @@ describe('NiftyswapExchange20', () => {
             { gasLimit: 8000000 }
           )
           if (condition[1] != 2) {
-            await expect(tx).to.be.rejectedWith(RevertError('SafeMath#sub: UNDERFLOW'))
+            await expect(tx).to.be.rejectedWith(RevertUnsafeMathError())
           } else {
             await expect(tx).to.be.rejectedWith(RevertError(erc1155_error_prefix + '_viewUpdateBinValue: UNDERFLOW'))
           }
         })
 
         it('should fail if token sent is 0', async () => {
-          let tokensAmountsToSellCopy = [...tokensAmountsToSell]
+          const tokensAmountsToSellCopy = [...tokensAmountsToSell]
           tokensAmountsToSellCopy[0] = BigNumber.from(0)
 
           const tx = userERC1155Contract.functions.safeBatchTransferFrom(
@@ -1731,9 +1712,15 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should fail if deadline is passed', async () => {
-          let timestamp = Math.floor(Date.now() / 1000) - 1
+          const timestamp = Math.floor(Date.now() / 1000) - 1
           const price = (await niftyswapExchangeContract.functions.getPrice_tokenToCurrency([0], [tokenAmountToSell]))[0]
-          let sellTokenData = getSellTokenData20(userAddress, price[0].mul(nTokenTypes), timestamp, extraFeeRecipients, extraFeeArray)
+          const sellTokenData = getSellTokenData20(
+            userAddress,
+            price[0].mul(nTokenTypes),
+            timestamp,
+            extraFeeRecipients,
+            extraFeeArray
+          )
 
           const tx = userERC1155Contract.functions.safeBatchTransferFrom(
             userAddress,
@@ -1748,7 +1735,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should pass if currency balance is equal to cost', async () => {
-          let sellTokenData = getSellTokenData20(userAddress, cost, deadline, extraFeeRecipients, extraFeeArray)
+          const sellTokenData = getSellTokenData20(userAddress, cost, deadline, extraFeeRecipients, extraFeeArray)
 
           const tx = userERC1155Contract.functions.safeBatchTransferFrom(
             userAddress,
@@ -1762,7 +1749,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should fail if currency balance is lower than cost', async () => {
-          let sellTokenData = getSellTokenData20(userAddress, cost.add(1), deadline, extraFeeRecipients, extraFeeArray)
+          const sellTokenData = getSellTokenData20(userAddress, cost.add(1), deadline, extraFeeRecipients, extraFeeArray)
 
           const tx = userERC1155Contract.functions.safeBatchTransferFrom(
             userAddress,
@@ -1784,9 +1771,7 @@ describe('NiftyswapExchange20', () => {
             sellTokenData,
             { gasLimit: 8000000 }
           )
-          await expect(tx1).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx1).to.be.rejectedWith(RevertError('NE20#32'))
 
           const tx2 = userERC1155Contract.functions.safeBatchTransferFrom(
             userAddress,
@@ -1796,9 +1781,7 @@ describe('NiftyswapExchange20', () => {
             sellTokenData,
             { gasLimit: 8000000 }
           )
-          await expect(tx2).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx2).to.be.rejectedWith(RevertError('NE20#32'))
 
           const tx3 = userERC1155Contract.functions.safeBatchTransferFrom(
             userAddress,
@@ -1808,9 +1791,7 @@ describe('NiftyswapExchange20', () => {
             sellTokenData,
             { gasLimit: 8000000 }
           )
-          await expect(tx3).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx3).to.be.rejectedWith(RevertError('NE20#32'))
         })
 
         it('should REVERT if arrays are not the same length', async () => {
@@ -1852,7 +1833,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should REVERT if not accounting for the extra fee', async () => {
-          sellTokenData = getSellTokenData20(userAddress, cost.add(extraFee), deadline, [extraFeeRecipient], [BigNumber.from(1)])
+          sellTokenData = getSellTokenData20(userAddress, cost.add(extraFee), deadline, [extraFeeRecipient], [One])
           const tx = userERC1155Contract.functions.safeBatchTransferFrom(
             userAddress,
             niftyswapExchangeContract.address,
@@ -1861,7 +1842,7 @@ describe('NiftyswapExchange20', () => {
             sellTokenData,
             { gasLimit: 8000000 }
           )
-          await expect(tx).to.be.rejectedWith(RevertError("NE20#8"))
+          await expect(tx).to.be.rejectedWith(RevertError('NE20#8'))
         })
 
         describe('When trade is successful', async () => {
@@ -1899,58 +1880,69 @@ describe('NiftyswapExchange20', () => {
             const reserves = await niftyswapExchangeContract.functions.getCurrencyReserves(types)
             for (let i = 0; i < types.length; i++) {
               // Extra fee and royalty aren't going to the reserve
-              expect(reserves[0][i]).to.be.eql(currencyAmountToAdd.sub((cost.add(extraFee).add(expectedRoyalty)).div(nTokenTypes)))
+              expect(reserves[0][i]).to.be.eql(currencyAmountToAdd.sub(cost.add(extraFee).add(expectedRoyalty).div(nTokenTypes)))
             }
           })
 
           it('should have token sell price adjusted', async () => {
-            const newCost: BigNumber = (await niftyswapExchangeContract.functions.getPrice_tokenToCurrency([0], [tokenAmountToSell]))[0][0]
-        
+            const newCost: BigNumber = (
+              await niftyswapExchangeContract.functions.getPrice_tokenToCurrency([0], [tokenAmountToSell])
+            )[0][0]
+
             const soldAmountWithFee = tokenAmountToSell.mul(LP_FEE_MULTIPLIER)
-            const currencyReserve = currencyAmountToAdd.sub((cost.add(extraFee).add(expectedRoyalty)).div(nTokenTypes))
-            let numerator = soldAmountWithFee.mul(currencyReserve)
-            let tokenReserveWithFee = tokenAmountToAdd.add(tokenAmountToSell).mul(1000)
-            let denominator = tokenReserveWithFee.add(soldAmountWithFee)
+            const currencyReserve = currencyAmountToAdd.sub(cost.add(extraFee).add(expectedRoyalty).div(nTokenTypes))
+            const numerator = soldAmountWithFee.mul(currencyReserve)
+            const tokenReserveWithFee = tokenAmountToAdd.add(tokenAmountToSell).mul(1000)
+            const denominator = tokenReserveWithFee.add(soldAmountWithFee)
             const newPreRoyaltyCost = numerator.div(denominator)
-            const newRoyalty = (newPreRoyaltyCost.mul(royaltyFee)).div(10000)
+            const newRoyalty = newPreRoyaltyCost.mul(royaltyFee).div(10000)
 
             expect(newCost).to.be.eql(newPreRoyaltyCost.sub(newRoyalty))
           })
 
           it('should emit CurrencyPurchase event', async () => {
-            let filterFromOperatorContract: ethers.ethers.EventFilter
-
             // Get event filter to get internal tx event
-            filterFromOperatorContract = niftyswapExchangeContract.filters.CurrencyPurchase(null, null, null, null, null, null, null)
+            const filterFromOperatorContract: EventFilter = niftyswapExchangeContract.filters.CurrencyPurchase(
+              null,
+              null,
+              null,
+              null,
+              null,
+              null,
+              null
+            )
 
             // Get logs from internal transaction event
-            // @ts-ignore (https://github.com/ethers-io/ethers.js/issues/204#issuecomment-427059031)
             filterFromOperatorContract.fromBlock = 0
-            let logs = await operatorProvider.getLogs(filterFromOperatorContract)
+            const logs = await operatorProvider.getLogs(filterFromOperatorContract)
             expect(logs[0].topics[0]).to.be.eql(
               niftyswapExchangeContract.interface.getEventTopic(
-                niftyswapExchangeContract.interface.events['CurrencyPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])']
+                niftyswapExchangeContract.interface.events[
+                  'CurrencyPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])'
+                ]
               )
             )
           })
 
           describe('CurrencyPurchase Event', () => {
-            let args;
+            let args
 
             beforeEach(async () => {
               const eventTopicHash = niftyswapExchangeContract.interface.getEventTopic(
-                niftyswapExchangeContract.interface.events['CurrencyPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])']
+                niftyswapExchangeContract.interface.events[
+                  'CurrencyPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])'
+                ]
               )
               const receipt = await tx.wait(1)
               const log = receipt.logs.find(a => a['topics'][0] === eventTopicHash)
               args = exchangeIface.parseLog(log).args
             })
-            
-            it('should have buyer address as `buyer` field', async () => {  
+
+            it('should have buyer address as `buyer` field', async () => {
               expect(args.buyer).to.be.eql(userAddress)
             })
-    
-            it('should have recipient address as `recipient` field', async () => {  
+
+            it('should have recipient address as `recipient` field', async () => {
               expect(args.recipient).to.be.eql(userAddress)
             })
 
@@ -1960,26 +1952,26 @@ describe('NiftyswapExchange20', () => {
               }
             })
 
-            it('should have tokensSoldAmounts as `tokensSoldAmounts` field', async () => {  
+            it('should have tokensSoldAmounts as `tokensSoldAmounts` field', async () => {
               for (let i = 0; i < types.length; i++) {
                 expect(args.tokensSoldAmounts[i]).to.be.eql(tokensAmountsToSell[i])
               }
             })
 
-            it('should have currencyBoughtAmounts as `currencyBoughtAmounts` field', async () => {  
-              const costPer = (cost.add(extraFee)).div(types.length)
+            it('should have currencyBoughtAmounts as `currencyBoughtAmounts` field', async () => {
+              const costPer = cost.add(extraFee).div(types.length)
               for (let i = 0; i < types.length; i++) {
                 expect(args.currencyBoughtAmounts[i]).to.be.eql(costPer)
               }
             })
 
-            it('should have extraFeeRecipients as `extraFeeRecipients` field', async () => {  
+            it('should have extraFeeRecipients as `extraFeeRecipients` field', async () => {
               for (let i = 0; i < types.length; i++) {
                 expect(args.extraFeeRecipients[i]).to.be.eql(extraFeeRecipients[i])
               }
             })
 
-            it('should have extraFeeAmounts as `extraFeeAmounts` field', async () => {  
+            it('should have extraFeeAmounts as `extraFeeAmounts` field', async () => {
               for (let i = 0; i < types.length; i++) {
                 expect(args.extraFeeAmounts[i]).to.be.eql(extraFeeArray[i])
               }
@@ -1987,10 +1979,9 @@ describe('NiftyswapExchange20', () => {
           })
 
           describe('Royalties Fees', () => {
-
             it('Royalty amount should be correct', async () => {
               const royalties = await niftyswapExchangeContract.functions.getRoyalties(randomAddress)
-              expect(royalties[0]).to.be.eql(expectedRoyalty)  
+              expect(royalties[0]).to.be.eql(expectedRoyalty)
               // Ignore extra fee
               expect(royalties[0]).to.be.eql(preRoyaltyCost.mul(nTokenTypes).sub(cost.add(extraFee)))
             })
@@ -2001,14 +1992,14 @@ describe('NiftyswapExchange20', () => {
             })
 
             it('Should allow anyone to send royalties to royalty recipient', async () => {
-              let tx = userExchangeContract.functions.sendRoyalties(randomAddress)
+              const tx = userExchangeContract.functions.sendRoyalties(randomAddress)
               await expect(tx).to.be.fulfilled
             })
-            
+
             context('when sendRoyalties() is successful', () => {
               let preWithdrawExchangeBalance
-              beforeEach(async function() {
-                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0] 
+              beforeEach(async function () {
+                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0]
                 await userExchangeContract.functions.sendRoyalties(randomAddress)
               })
 
@@ -2032,7 +2023,7 @@ describe('NiftyswapExchange20', () => {
           describe('Extra Fees', () => {
             it('Extra amount should be correct', async () => {
               const royalties = await niftyswapExchangeContract.functions.getRoyalties(extraFeeRecipient)
-              expect(royalties[0]).to.be.eql(extraFee)  
+              expect(royalties[0]).to.be.eql(extraFee)
             })
 
             it('Should allow royalty recipient to withdraw', async () => {
@@ -2041,14 +2032,14 @@ describe('NiftyswapExchange20', () => {
             })
 
             it('Should allow anyone to send royalties to royalty recipient', async () => {
-              let tx = userExchangeContract.functions.sendRoyalties(extraFeeRecipient)
+              const tx = userExchangeContract.functions.sendRoyalties(extraFeeRecipient)
               await expect(tx).to.be.fulfilled
             })
-            
+
             context('when sendRoyalties() is successful', () => {
               let preWithdrawExchangeBalance
-              beforeEach(async function() {
-                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0] 
+              beforeEach(async function () {
+                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0]
                 await userExchangeContract.functions.sendRoyalties(extraFeeRecipient)
               })
 
@@ -2064,18 +2055,17 @@ describe('NiftyswapExchange20', () => {
 
               it('Should set claimable royalty to 0', async () => {
                 const royalty = await userExchangeContract.functions.getRoyalties(extraFeeRecipient)
-                await expect(royalty[0]).to.be.eql(BigNumber.from(0))
+                expect(royalty[0]).to.be.eql(BigNumber.from(0))
               })
             })
           })
-
         })
       })
 
       describe('_currencyToToken() function', () => {
         //Buy
         const tokenAmountToBuy = BigNumber.from(50)
-        const tokensAmountsToBuy: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => tokenAmountToBuy)
+        const tokensAmountsToBuy: BigNumber[] = new Array(nTokenTypes).fill('').map(() => tokenAmountToBuy)
         let cost: BigNumber
         let expectedRoyalty: BigNumber
         let preRoyaltyCost: BigNumber
@@ -2092,13 +2082,15 @@ describe('NiftyswapExchange20', () => {
           )
 
           // Sell
-          const allcosts = (await niftyswapExchangeContract.functions.getPrice_currencyToToken(types, tokensAmountsToBuy))
+          const allcosts = await niftyswapExchangeContract.functions.getPrice_currencyToToken(types, tokensAmountsToBuy)
           cost = allcosts[0].reduce((acc: BigNumber, a: BigNumber) => acc.add(a)).add(extraFee)
 
-          // Expected royalty 
-          const tokenReserve = (await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, 0))
+          // Expected royalty
+          const tokenReserve = await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, 0)
           const currencyReserve = (await niftyswapExchangeContract.functions.getCurrencyReserves([0]))[0][0]
-          preRoyaltyCost = (await niftyswapExchangeContract.functions.getBuyPrice(tokenAmountToBuy, currencyReserve, tokenReserve))[0]
+          preRoyaltyCost = (
+            await niftyswapExchangeContract.functions.getBuyPrice(tokenAmountToBuy, currencyReserve, tokenReserve)
+          )[0]
           expectedRoyalty = preRoyaltyCost.mul(royaltyFee).div(10000).mul(nTokenTypes)
         })
 
@@ -2118,7 +2110,7 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should fail if deadline is passed', async () => {
-          let timestamp = Math.floor(Date.now() / 1000) - 1
+          const timestamp = Math.floor(Date.now() / 1000) - 1
           const tx = userExchangeContract.functions.buyTokens(
             types,
             tokensAmountsToBuy,
@@ -2133,8 +2125,6 @@ describe('NiftyswapExchange20', () => {
         })
 
         it('should fail if any duplicate', async () => {
-          let One = BigNumber.from(1)
-
           // Tokens to buy
           const tx1 = userExchangeContract.functions.buyTokens(
             [1, 1],
@@ -2146,9 +2136,7 @@ describe('NiftyswapExchange20', () => {
             extraFeeArray,
             { gasLimit: 8000000 }
           )
-          await expect(tx1).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx1).to.be.rejectedWith(RevertError('NE20#32'))
 
           // Tokens to buy
           const tx2 = userExchangeContract.functions.buyTokens(
@@ -2161,9 +2149,7 @@ describe('NiftyswapExchange20', () => {
             extraFeeArray,
             { gasLimit: 8000000 }
           )
-          await expect(tx2).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx2).to.be.rejectedWith(RevertError('NE20#32'))
 
           // Tokens to buy
           const tx3 = userExchangeContract.functions.buyTokens(
@@ -2176,9 +2162,7 @@ describe('NiftyswapExchange20', () => {
             extraFeeArray,
             { gasLimit: 8000000 }
           )
-          await expect(tx3).to.be.rejectedWith(
-            RevertError('NE20#32')
-          )
+          await expect(tx3).to.be.rejectedWith(RevertError('NE20#32'))
         })
 
         it('should REVERT if arrays are not the same length', async () => {
@@ -2192,7 +2176,7 @@ describe('NiftyswapExchange20', () => {
             extraFeeArray,
             { gasLimit: 8000000 }
           )
-          await expect(tx1).to.be.rejectedWith(OpCodeError())
+          await expect(tx1).to.be.rejectedWith(ArrayAccessError())
 
           const tx2 = userExchangeContract.functions.buyTokens(
             [0],
@@ -2231,14 +2215,14 @@ describe('NiftyswapExchange20', () => {
             deadline,
             userAddress,
             [extraFeeRecipient],
-            [BigNumber.from(1)],
+            [One],
             { gasLimit: 8000000 }
           )
-          await expect(tx).to.be.rejectedWith(RevertError("SafeMath#sub: UNDERFLOW"))
+          await expect(tx).to.be.rejectedWith(RevertUnsafeMathError())
         })
 
         describe('When trade is successful', async () => {
-          let tx; 
+          let tx
 
           beforeEach(async () => {
             tx = await userExchangeContract.functions.buyTokens(
@@ -2274,58 +2258,67 @@ describe('NiftyswapExchange20', () => {
           it('should update the currency per token reserve', async () => {
             const reserves = await niftyswapExchangeContract.functions.getCurrencyReserves(types)
             for (let i = 0; i < types.length; i++) {
-              expect(reserves[0][i]).to.be.eql(currencyAmountToAdd.add((cost.sub(expectedRoyalty).sub(extraFee)).div(nTokenTypes)))
+              expect(reserves[0][i]).to.be.eql(currencyAmountToAdd.add(cost.sub(expectedRoyalty).sub(extraFee).div(nTokenTypes)))
             }
           })
 
           it('should have token buy price adjusted', async () => {
             const newCost = (await niftyswapExchangeContract.functions.getPrice_currencyToToken([0], [tokenAmountToBuy]))[0]
 
-            let currencyReserve = currencyAmountToAdd.add((cost.sub(expectedRoyalty).sub(extraFee)).div(nTokenTypes))
-            let tokenReserve = tokenAmountToAdd.sub(tokenAmountToBuy)
+            const currencyReserve = currencyAmountToAdd.add(cost.sub(expectedRoyalty).sub(extraFee).div(nTokenTypes))
+            const tokenReserve = tokenAmountToAdd.sub(tokenAmountToBuy)
 
-            let numerator = currencyReserve.mul(tokenAmountToBuy).mul(1000)
-            let denominator = tokenReserve.sub(tokenAmountToBuy).mul(LP_FEE_MULTIPLIER)
+            const numerator = currencyReserve.mul(tokenAmountToBuy).mul(1000)
+            const denominator = tokenReserve.sub(tokenAmountToBuy).mul(LP_FEE_MULTIPLIER)
             const newPreRoyaltyCost = numerator.div(denominator).add(1)
-            const newRoyalty = (newPreRoyaltyCost.mul(royaltyFee)).div(10000)
+            const newRoyalty = newPreRoyaltyCost.mul(royaltyFee).div(10000)
 
             expect(newCost[0]).to.be.eql(newPreRoyaltyCost.add(newRoyalty))
           })
 
           it('should emit TokensPurchase event', async () => {
-            let filterFromOperatorContract: ethers.ethers.EventFilter
-
             // Get event filter to get internal tx event
-            filterFromOperatorContract = niftyswapExchangeContract.filters.TokensPurchase(null, null, null, null, null, null, null)
+            const filterFromOperatorContract: EventFilter = niftyswapExchangeContract.filters.TokensPurchase(
+              null,
+              null,
+              null,
+              null,
+              null,
+              null,
+              null
+            )
 
             // Get logs from internal transaction event
-            // @ts-ignore (https://github.com/ethers-io/ethers.js/issues/204#issuecomment-427059031)
             filterFromOperatorContract.fromBlock = 0
-            let logs = await operatorProvider.getLogs(filterFromOperatorContract)
+            const logs = await operatorProvider.getLogs(filterFromOperatorContract)
             expect(logs[0].topics[0]).to.be.eql(
               niftyswapExchangeContract.interface.getEventTopic(
-                niftyswapExchangeContract.interface.events['TokensPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])']
+                niftyswapExchangeContract.interface.events[
+                  'TokensPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])'
+                ]
               )
             )
           })
 
           describe('TokensPurchase Event', () => {
-            let args;
+            let args
 
             beforeEach(async () => {
               const eventTopicHash = niftyswapExchangeContract.interface.getEventTopic(
-                niftyswapExchangeContract.interface.events['TokensPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])']
+                niftyswapExchangeContract.interface.events[
+                  'TokensPurchase(address,address,uint256[],uint256[],uint256[],address[],uint256[])'
+                ]
               )
               const receipt = await tx.wait(1)
               const log = receipt.logs.find(a => a['topics'][0] === eventTopicHash)
               args = exchangeIface.parseLog(log).args
             })
-            
-            it('should have buyer address as `buyer` field', async () => {  
+
+            it('should have buyer address as `buyer` field', async () => {
               expect(args.buyer).to.be.eql(userAddress)
             })
-    
-            it('should have recipient address as `recipient` field', async () => {  
+
+            it('should have recipient address as `recipient` field', async () => {
               expect(args.recipient).to.be.eql(userAddress)
             })
 
@@ -2335,26 +2328,26 @@ describe('NiftyswapExchange20', () => {
               }
             })
 
-            it('should have tokensBoughtAmounts as `tokensBoughtAmounts` field', async () => {  
+            it('should have tokensBoughtAmounts as `tokensBoughtAmounts` field', async () => {
               for (let i = 0; i < types.length; i++) {
                 expect(args.tokensBoughtAmounts[i]).to.be.eql(tokensAmountsToBuy[i])
               }
             })
 
-            it('should have currencySoldAmounts as `currencySoldAmounts` field', async () => {  
-              const costPer = (cost.sub(extraFee)).div(types.length)
+            it('should have currencySoldAmounts as `currencySoldAmounts` field', async () => {
+              const costPer = cost.sub(extraFee).div(types.length)
               for (let i = 0; i < types.length; i++) {
                 expect(args.currencySoldAmounts[i]).to.be.eql(costPer)
               }
             })
 
-            it('should have extraFeeRecipients as `extraFeeRecipients` field', async () => {  
+            it('should have extraFeeRecipients as `extraFeeRecipients` field', async () => {
               for (let i = 0; i < types.length; i++) {
                 expect(args.extraFeeRecipients[i]).to.be.eql(extraFeeRecipients[i])
               }
             })
 
-            it('should have extraFeeAmounts as `extraFeeAmounts` field', async () => {  
+            it('should have extraFeeAmounts as `extraFeeAmounts` field', async () => {
               for (let i = 0; i < types.length; i++) {
                 expect(args.extraFeeAmounts[i]).to.be.eql(extraFeeArray[i])
               }
@@ -2364,7 +2357,7 @@ describe('NiftyswapExchange20', () => {
           describe('Royalties Fees', () => {
             it('Royalty amount should be correct', async () => {
               const royalties = await niftyswapExchangeContract.functions.getRoyalties(randomAddress)
-              expect(royalties[0]).to.be.eql(expectedRoyalty)  
+              expect(royalties[0]).to.be.eql(expectedRoyalty)
               expect(royalties[0]).to.be.eql(cost.sub(preRoyaltyCost.mul(nTokenTypes)).sub(extraFee))
             })
 
@@ -2374,14 +2367,14 @@ describe('NiftyswapExchange20', () => {
             })
 
             it('Should allow anyone to send royalties to royalty recipient', async () => {
-              let tx = userExchangeContract.functions.sendRoyalties(randomAddress)
+              const tx = userExchangeContract.functions.sendRoyalties(randomAddress)
               await expect(tx).to.be.fulfilled
             })
-            
+
             context('when sendRoyalties() is successful', () => {
               let preWithdrawExchangeBalance
-              beforeEach(async function() {
-                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0] 
+              beforeEach(async function () {
+                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0]
                 await userExchangeContract.functions.sendRoyalties(randomAddress)
               })
 
@@ -2405,7 +2398,7 @@ describe('NiftyswapExchange20', () => {
           describe('Extra Fees', () => {
             it('Extra amount should be correct', async () => {
               const royalties = await niftyswapExchangeContract.functions.getRoyalties(extraFeeRecipient)
-              expect(royalties[0]).to.be.eql(extraFee)  
+              expect(royalties[0]).to.be.eql(extraFee)
             })
 
             it('Should allow royalty recipient to withdraw', async () => {
@@ -2414,14 +2407,14 @@ describe('NiftyswapExchange20', () => {
             })
 
             it('Should allow anyone to send royalties to royalty recipient', async () => {
-              let tx = userExchangeContract.functions.sendRoyalties(extraFeeRecipient)
+              const tx = userExchangeContract.functions.sendRoyalties(extraFeeRecipient)
               await expect(tx).to.be.fulfilled
             })
-            
+
             context('when sendRoyalties() is successful', () => {
               let preWithdrawExchangeBalance
-              beforeEach(async function() {
-                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0] 
+              beforeEach(async function () {
+                preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0]
                 await userExchangeContract.functions.sendRoyalties(extraFeeRecipient)
               })
 
@@ -2469,9 +2462,7 @@ describe('NiftyswapExchange20', () => {
             expect(userBalance[0]).to.be.eql(BigNumber.from(nTokensPerType))
           }
 
-          const exchangeCurrencyBalance = await userCurrencyContract.functions.balanceOf(
-            niftyswapExchangeContract.address,
-          )
+          const exchangeCurrencyBalance = await userCurrencyContract.functions.balanceOf(niftyswapExchangeContract.address)
           const randomCurrencyBalance = await userCurrencyContract.functions.balanceOf(randomWallet.address)
           const userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
 
@@ -2504,9 +2495,7 @@ describe('NiftyswapExchange20', () => {
             expect(userBalance[0]).to.be.eql(BigNumber.from(nTokensPerType).add(tokenAmountToBuy))
           }
 
-          const exchangeCurrencyBalance = await userCurrencyContract.functions.balanceOf(
-            niftyswapExchangeContract.address,
-          )
+          const exchangeCurrencyBalance = await userCurrencyContract.functions.balanceOf(niftyswapExchangeContract.address)
           const userCurrencyBalance = await userCurrencyContract.functions.balanceOf(userAddress)
 
           expect(exchangeCurrencyBalance[0]).to.be.eql(currencyAmountToAdd.mul(nTokenTypes).add(cost))
@@ -2514,11 +2503,13 @@ describe('NiftyswapExchange20', () => {
         })
       })
 
-      describe('ERC-2981 Token', () => {       
+      describe('ERC-2981 Token', () => {
         // Only run for ERC-2981 tokens
-        before(async function() {
+        before(async function () {
           if (!(condition[1] == 3 || condition[1] == 4)) {
-            this.test!.parent!.pending = true
+            if (this.test?.parent) {
+              this.test.parent.pending = true
+            }
             this.skip()
           }
         })
@@ -2548,27 +2539,32 @@ describe('NiftyswapExchange20', () => {
           const tokenAmountToBuy = BigNumber.from(10)
           const tokenAmountToAdd = BigNumber.from(50)
           const idsTobuy: number[] = [666, 777, 888]
-          const tokensAmountsToMint: BigNumber[] = new Array(idsTobuy.length).fill('').map((a, i) => tokenAmountToBuy.add(tokenAmountToAdd))
-          const tokensAmountsToAdd: BigNumber[] = new Array(idsTobuy.length).fill('').map((a, i) => tokenAmountToAdd)
-          const tokensAmountsToBuy: BigNumber[] = new Array(idsTobuy.length).fill('').map((a, i) => tokenAmountToBuy)
+          const tokensAmountsToMint: BigNumber[] = new Array(idsTobuy.length)
+            .fill('')
+            .map(() => tokenAmountToBuy.add(tokenAmountToAdd))
+          const tokensAmountsToAdd: BigNumber[] = new Array(idsTobuy.length).fill('').map(() => tokenAmountToAdd)
+          const tokensAmountsToBuy: BigNumber[] = new Array(idsTobuy.length).fill('').map(() => tokenAmountToBuy)
 
           const recipient1 = ethers.Wallet.createRandom().address
           const recipient666 = ethers.Wallet.createRandom().address
 
           const royaltyFee = 200
           const royaltyFee666 = 1000
-          
+
           let expectedRoyalty: BigNumber
           let expectedRoyalty666: BigNumber
           let cost
-          
+
           beforeEach(async () => {
             await ownerERC1155Contract.functions.batchMintMock(operatorAddress, idsTobuy, tokensAmountsToMint, [])
             await ownerERC1155Contract.functions.batchMintMock(userAddress, idsTobuy, tokensAmountsToMint, [])
 
-            const addLiquidityData: string = getAddLiquidityData([currencyAmountToAdd, currencyAmountToAdd, currencyAmountToAdd], deadline)
+            const addLiquidityData: string = getAddLiquidityData(
+              [currencyAmountToAdd, currencyAmountToAdd, currencyAmountToAdd],
+              deadline
+            )
 
-            // Add liquidity 
+            // Add liquidity
             await operatorERC1155Contract.functions.safeBatchTransferFrom(
               operatorAddress,
               niftyswapExchangeContract.address,
@@ -2585,15 +2581,17 @@ describe('NiftyswapExchange20', () => {
             await (ownerERC1155Contract as ERC1155RoyaltyMock).functions.set666FeeRecipient(recipient666)
 
             // Calculate expected royalties
-            const token_reserve = (await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, idsTobuy[0]))
+            const token_reserve = await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, idsTobuy[0])
             const currency_reserve = (await niftyswapExchangeContract.functions.getCurrencyReserves([idsTobuy[0]]))[0][0]
-            const preRoyaltyCost = (await niftyswapExchangeContract.functions.getBuyPrice(tokenAmountToBuy, currency_reserve, token_reserve))[0]
+            const preRoyaltyCost = (
+              await niftyswapExchangeContract.functions.getBuyPrice(tokenAmountToBuy, currency_reserve, token_reserve)
+            )[0]
 
             expectedRoyalty = preRoyaltyCost.mul(royaltyFee).div(10000).mul(2)
             expectedRoyalty666 = preRoyaltyCost.mul(royaltyFee666).div(10000)
-            
+
             // Calculate cost
-            const allcosts = (await niftyswapExchangeContract.functions.getPrice_currencyToToken(idsTobuy, tokensAmountsToBuy))
+            const allcosts = await niftyswapExchangeContract.functions.getPrice_currencyToToken(idsTobuy, tokensAmountsToBuy)
             cost = allcosts[0].reduce((acc: BigNumber, a: BigNumber) => acc.add(a)).add(extraFee)
 
             // Buy tokens
@@ -2611,21 +2609,21 @@ describe('NiftyswapExchange20', () => {
 
           it('Royalty amount for first recipient should be correct', async () => {
             const royalties = await niftyswapExchangeContract.functions.getRoyalties(recipient1)
-            expect(royalties[0]).to.be.eql(expectedRoyalty)  
+            expect(royalties[0]).to.be.eql(expectedRoyalty)
           })
 
           it('Royalty amount for 2nd recipient should be correct', async () => {
             const royalties666 = await niftyswapExchangeContract.functions.getRoyalties(recipient666)
-            expect(royalties666[0]).to.be.eql(expectedRoyalty666)  
+            expect(royalties666[0]).to.be.eql(expectedRoyalty666)
 
             const royalties = await niftyswapExchangeContract.functions.getRoyalties(recipient1)
-            expect(royalties[0].div(2)).to.be.eql(royalties666[0].div(royaltyFee666 / royaltyFee))  
+            expect(royalties[0].div(2)).to.be.eql(royalties666[0].div(royaltyFee666 / royaltyFee))
           })
-                    
+
           context('when sendRoyalties() is successful', () => {
             let preWithdrawExchangeBalance
-            beforeEach(async function() {
-              preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0] 
+            beforeEach(async function () {
+              preWithdrawExchangeBalance = (await ownerCurrencyContract.functions.balanceOf(userExchangeContract.address))[0]
               await userExchangeContract.functions.sendRoyalties(recipient1)
               await userExchangeContract.functions.sendRoyalties(recipient666)
             })
@@ -2654,9 +2652,9 @@ describe('NiftyswapExchange20', () => {
 
       describe('Edge cases', () => {
         it('Pool can not go to zero token in reserve', async () => {
-          const minBaseCurrency = BigNumber.from(10**3)
-          const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => minBaseCurrency)
-          const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => BigNumber.from(1))
+          const minBaseCurrency = BigNumber.from(10 ** 3)
+          const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => minBaseCurrency)
+          const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => One)
           const addLiquidityData: string = getAddLiquidityData(currencyAmountsToAdd, deadline)
 
           // Add 1000:1 for all pools
@@ -2670,14 +2668,14 @@ describe('NiftyswapExchange20', () => {
           )
 
           // Trying to buy the only tokn will fail as it will cause a division by 0
-          let tx = niftyswapExchangeContract.functions.getPrice_currencyToToken([types[0]], [1])
-          await expect(tx).to.be.rejectedWith(OpCodeError())
+          const tx = niftyswapExchangeContract.functions.getPrice_currencyToToken([types[0]], [1])
+          await expect(tx).to.be.rejectedWith(CallError())
         })
 
         it('Pool stuck at 1 token can go back up to normal with loss', async () => {
           const initialBaseCurrency = BigNumber.from(10 ** 9)
-          const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => initialBaseCurrency)
-          const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => BigNumber.from(1))
+          const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => initialBaseCurrency)
+          const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => One)
           const addLiquidityData: string = getAddLiquidityData(currencyAmountsToAdd, deadline)
 
           // Add 1000:1 for all pools
@@ -2700,9 +2698,9 @@ describe('NiftyswapExchange20', () => {
           // which is the desired price for users to start selling the assets or add liquidity.
 
           // Add 31622 tokens to pool
-          let maxBaseCurrency_1 = BigNumber.from(10).pow(18)
-          let currencyAmountsToAdd_1 = new Array(nTokenTypes).fill('').map((a, i) => maxBaseCurrency_1)
-          const tokenAmountsToAdd_1 = new Array(nTokenTypes).fill('').map((a, i) => BigNumber.from(31622))
+          const maxBaseCurrency_1 = BigNumber.from(10).pow(18)
+          const currencyAmountsToAdd_1 = new Array(nTokenTypes).fill('').map(() => maxBaseCurrency_1)
+          const tokenAmountsToAdd_1 = new Array(nTokenTypes).fill('').map(() => BigNumber.from(31622))
           const addLiquidityData_1 = getAddLiquidityData(currencyAmountsToAdd_1, deadline)
           tx = operatorERC1155Contract.functions.safeBatchTransferFrom(
             operatorAddress,
@@ -2715,13 +2713,17 @@ describe('NiftyswapExchange20', () => {
           await expect(tx).to.be.fulfilled
 
           // Buy 31622 tokens, to leave a ratio of 10**18 : 1, testing with 1 pool
-          let amount_to_buy = BigNumber.from(31622)
-          let cost = (await niftyswapExchangeContract.functions.getPrice_currencyToToken([types[0]], [amount_to_buy]))[0][0].add(extraFee)
-          
-          // Expected royalty 
-          const token_reserve = (await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, types[0]))
+          const amount_to_buy = BigNumber.from(31622)
+          const cost = (
+            await niftyswapExchangeContract.functions.getPrice_currencyToToken([types[0]], [amount_to_buy])
+          )[0][0].add(extraFee)
+
+          // Expected royalty
+          const token_reserve = await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, types[0])
           const currency_reserve = (await niftyswapExchangeContract.functions.getCurrencyReserves([types[0]]))[0][0]
-          const preRoyaltyCost = (await niftyswapExchangeContract.functions.getBuyPrice(amount_to_buy, currency_reserve, token_reserve))[0]
+          const preRoyaltyCost = (
+            await niftyswapExchangeContract.functions.getBuyPrice(amount_to_buy, currency_reserve, token_reserve)
+          )[0]
           const expectedRoyalty = preRoyaltyCost.mul(royaltyFee).div(10000)
 
           // Perform the puirchase
@@ -2737,12 +2739,12 @@ describe('NiftyswapExchange20', () => {
           )
           await expect(tx).to.be.fulfilled
 
-          const new_token_reserve = (await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, types[0]))
+          const new_token_reserve = await ownerERC1155Contract.balanceOf(niftyswapExchangeContract.address, types[0])
           const new_currency_reserve = (await niftyswapExchangeContract.functions.getCurrencyReserves([types[0]]))[0][0]
 
           // Pool should have more than 10**18 currency and 1 token
-          let expected_price = cost.add(amount_to_buy.add(1).mul(BigNumber.from(10).pow(9)))
-          expect(new_token_reserve).to.be.eql(BigNumber.from(1))
+          const expected_price = cost.add(amount_to_buy.add(1).mul(BigNumber.from(10).pow(9)))
+          expect(new_token_reserve).to.be.eql(One)
           expect(new_currency_reserve).to.be.eql(expected_price.sub(expectedRoyalty).sub(extraFee))
         })
       })
