@@ -7,17 +7,12 @@ import {
   getBuyTokenData,
   getSellTokenData,
   getAddLiquidityData,
-  getRemoveLiquidityData
+  HIGH_GAS_LIMIT
 } from './utils'
 
 import * as utils from './utils'
 
-import {
-  ERC1155Mock,
-  NiftyswapExchange,
-  NiftyswapFactory,
-  WrapAndNiftyswap
-} from 'src/gen/typechain'
+import { ERC1155Mock, NiftyswapExchange, NiftyswapFactory, WrapAndNiftyswap } from 'src/gen/typechain'
 
 import { ERC20Mock, ERC20Wrapper } from '@0xsequence/erc20-meta-token'
 
@@ -27,13 +22,9 @@ import { web3 } from 'hardhat'
 
 // init test wallets from package.json mnemonic
 
-const { wallet: ownerWallet, provider: ownerProvider, signer: ownerSigner } = utils.createTestWallet(web3, 0)
-
-const { wallet: userWallet, provider: userProvider, signer: userSigner } = utils.createTestWallet(web3, 2)
-
-const { wallet: operatorWallet, provider: operatorProvider, signer: operatorSigner } = utils.createTestWallet(web3, 4)
-
-const { wallet: randomWallet, provider: randomProvider, signer: randomSigner } = utils.createTestWallet(web3, 5)
+const { wallet: ownerWallet, provider: ownerProvider, signer: ownerSigner } = utils.createTestWallet(web3, 0) // eslint-disable-line @typescript-eslint/no-unused-vars
+const { wallet: userWallet, provider: userProvider, signer: userSigner } = utils.createTestWallet(web3, 2) // eslint-disable-line @typescript-eslint/no-unused-vars
+const { wallet: operatorWallet, provider: operatorProvider, signer: operatorSigner } = utils.createTestWallet(web3, 4) // eslint-disable-line @typescript-eslint/no-unused-vars
 
 const getBig = (id: number) => BigNumber.from(id)
 
@@ -79,7 +70,7 @@ describe('WrapAndSwap', () => {
   let wrapAndSwap: string
 
   // Token Param
-  const nTokenTypes = 30 //560
+  const nTokenTypes = 3 //560
   const nTokensPerType = 500000
 
   // Currency Param
@@ -88,9 +79,7 @@ describe('WrapAndSwap', () => {
 
   // Add liquidity data
   const tokenAmountToAdd = BigNumber.from(300)
-  const currencyAmountToAdd = BigNumber.from(10)
-    .pow(18)
-    .mul(299)
+  const currencyAmountToAdd = BigNumber.from(10).pow(18).mul(299)
 
   // Transactions parameters
   const TX_PARAM = { gasLimit: 5000000 }
@@ -98,10 +87,10 @@ describe('WrapAndSwap', () => {
   const deadline = Math.floor(Date.now() / 1000) + 100000
 
   // Arrays
-  const types = new Array(nTokenTypes).fill('').map((a, i) => getBig(i))
-  const values = new Array(nTokenTypes).fill('').map((a, i) => nTokensPerType)
-  const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => currencyAmountToAdd)
-  const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => tokenAmountToAdd)
+  const types = new Array(nTokenTypes).fill('').map((_a, i) => getBig(i))
+  const values = new Array(nTokenTypes).fill('').map(() => nTokensPerType)
+  const currencyAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => currencyAmountToAdd)
+  const tokenAmountsToAdd: BigNumber[] = new Array(nTokenTypes).fill('').map(() => tokenAmountToAdd)
   const addLiquidityData: string = getAddLiquidityData(currencyAmountsToAdd, deadline)
 
   // load contract abi and deploy to test server
@@ -180,12 +169,7 @@ describe('WrapAndSwap', () => {
     await ownerERC20Contract.functions.mockMint(userAddress, currencyAmount)
 
     // Wrap some tokens for niftyswap liquidity
-    await operatorERC20Contract.functions.approve(
-      ownerTokenWrapper.address,
-      BigNumber.from(2)
-        .pow(256)
-        .sub(1)
-    )
+    await operatorERC20Contract.functions.approve(ownerTokenWrapper.address, BigNumber.from(2).pow(256).sub(1))
     await operatorTokenWrapper.functions.deposit(
       operatorERC20Contract.address,
       operatorAddress,
@@ -207,18 +191,12 @@ describe('WrapAndSwap', () => {
     )
 
     // User approves wrapAndSwap
-    await userERC20Contract.functions.approve(
-      userWrapAndNiftyswap.address,
-      BigNumber.from(2)
-        .pow(256)
-        .sub(1),
-      TX_PARAM
-    )
+    await userERC20Contract.functions.approve(userWrapAndNiftyswap.address, BigNumber.from(2).pow(256).sub(1), TX_PARAM)
   })
 
   describe('wrapAndSwap() function', () => {
     const tokenAmountToBuy = BigNumber.from(50)
-    const tokensAmountsToBuy: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => tokenAmountToBuy)
+    const tokensAmountsToBuy: BigNumber[] = new Array(nTokenTypes).fill('').map(() => tokenAmountToBuy)
     let buyTokenData: string
     let cost: BigNumber
 
@@ -229,28 +207,28 @@ describe('WrapAndSwap', () => {
     })
 
     it('should revert if order recipient is not swapAndWrap contract', async () => {
-      let bad_buyTokenData = getBuyTokenData(userAddress, types, tokensAmountsToBuy, deadline)
-      const tx = userWrapAndNiftyswap.functions.wrapAndSwap(cost, userAddress, bad_buyTokenData, { gasLimit: 10000000 })
+      const bad_buyTokenData = getBuyTokenData(userAddress, types, tokensAmountsToBuy, deadline)
+      const tx = userWrapAndNiftyswap.functions.wrapAndSwap(cost, userAddress, bad_buyTokenData, HIGH_GAS_LIMIT)
       await expect(tx).to.be.rejectedWith(RevertError('WrapAndNiftyswap#wrapAndSwap: ORDER RECIPIENT MUST BE THIS CONTRACT'))
     })
 
     it('should buy tokens when balances are sufficient', async () => {
-      const tx = userWrapAndNiftyswap.functions.wrapAndSwap(cost, userAddress, buyTokenData, { gasLimit: 10000000 })
+      const tx = userWrapAndNiftyswap.functions.wrapAndSwap(cost, userAddress, buyTokenData, HIGH_GAS_LIMIT)
       await expect(tx).to.be.fulfilled
     })
 
     it('should buy the 2nd time as well', async () => {
-      await userWrapAndNiftyswap.functions.wrapAndSwap(cost, userAddress, buyTokenData, { gasLimit: 10000000 })
+      await userWrapAndNiftyswap.functions.wrapAndSwap(cost, userAddress, buyTokenData, HIGH_GAS_LIMIT)
       let cost2 = (await niftyswapExchangeContract.functions.getPrice_currencyToToken([0], [tokenAmountToBuy]))[0][0]
-      cost2 = cost.mul(nTokenTypes)
-      let buyTokenData2 = getBuyTokenData(ZERO_ADDRESS, types, tokensAmountsToBuy, deadline)
-      let tx = userWrapAndNiftyswap.functions.wrapAndSwap(cost2, userAddress, buyTokenData2, { gasLimit: 10000000 })
+      cost2 = cost2.mul(nTokenTypes)
+      const buyTokenData2 = getBuyTokenData(ZERO_ADDRESS, types, tokensAmountsToBuy, deadline)
+      const tx = userWrapAndNiftyswap.functions.wrapAndSwap(cost2, userAddress, buyTokenData2, HIGH_GAS_LIMIT)
       await expect(tx).to.be.fulfilled
     })
 
     context('When wrapAndSwap is completed', () => {
       beforeEach(async () => {
-        await userWrapAndNiftyswap.functions.wrapAndSwap(cost.add(100), userAddress, buyTokenData, { gasLimit: 10000000 })
+        await userWrapAndNiftyswap.functions.wrapAndSwap(cost.add(100), userAddress, buyTokenData, HIGH_GAS_LIMIT)
       })
 
       it('should update Tokens balances if it passes', async () => {
@@ -275,7 +253,7 @@ describe('WrapAndSwap', () => {
         const erc20Balance = await userERC20Contract.functions.balanceOf(userWrapAndNiftyswap.address)
         const wrappedTokenBalance = await userTokenWrapper.functions.balanceOf(userWrapAndNiftyswap.address, currencyID)
 
-        let addresses = new Array(nTokenTypes).fill('').map((a, i) => userWrapAndNiftyswap.address)
+        const addresses = new Array(nTokenTypes).fill('').map(() => userWrapAndNiftyswap.address)
         const erc1155Balances = await userERC1155Contract.functions.balanceOfBatch(addresses, types)
 
         expect(erc20Balance[0]).to.be.eql(ethers.constants.Zero)
@@ -289,9 +267,9 @@ describe('WrapAndSwap', () => {
 
   describe('swapAndUnwrap() function', () => {
     const tokenAmountToSell = BigNumber.from(50)
-    const tokensAmountsToSell: BigNumber[] = new Array(nTokenTypes).fill('').map((a, i) => tokenAmountToSell)
+    const tokensAmountsToSell: BigNumber[] = new Array(nTokenTypes).fill('').map(() => tokenAmountToSell)
     let sellTokenData: string
-    let expectedAmount
+    let expectedAmount: BigNumber
 
     beforeEach(async () => {
       // Sell
@@ -301,14 +279,14 @@ describe('WrapAndSwap', () => {
     })
 
     it('should revert if order recipient is not swapAndWrap contract', async () => {
-      let bad_sellTokenData = getSellTokenData(userAddress, expectedAmount, deadline)
+      const bad_sellTokenData = getSellTokenData(userAddress, expectedAmount, deadline)
       const tx = userERC1155Contract.functions.safeBatchTransferFrom(
         userAddress,
         wrapAndSwap,
         types,
         tokensAmountsToSell,
         bad_sellTokenData,
-        { gasLimit: 10000000 }
+        HIGH_GAS_LIMIT
       )
       await expect(tx).to.be.rejectedWith(
         RevertError('WrapAndNiftyswap#onERC1155BatchReceived: ORDER RECIPIENT MUST BE THIS CONTRACT')
@@ -336,10 +314,10 @@ describe('WrapAndSwap', () => {
         sellTokenData,
         TX_PARAM
       )
-      let price2 = (await niftyswapExchangeContract.functions.getPrice_tokenToCurrency([0], [tokenAmountToSell]))[0]
-      let expectedAmount2 = price2[0].mul(nTokenTypes)
-      let sellTokenData2 = getSellTokenData(ZERO_ADDRESS, expectedAmount2, deadline)
-      let tx = userERC1155Contract.functions.safeBatchTransferFrom(
+      const price2 = (await niftyswapExchangeContract.functions.getPrice_tokenToCurrency([0], [tokenAmountToSell]))[0]
+      const expectedAmount2 = price2[0].mul(nTokenTypes)
+      const sellTokenData2 = getSellTokenData(ZERO_ADDRESS, expectedAmount2, deadline)
+      const tx = userERC1155Contract.functions.safeBatchTransferFrom(
         userAddress,
         wrapAndSwap,
         types,
@@ -384,7 +362,7 @@ describe('WrapAndSwap', () => {
         const erc20Balance = await userERC20Contract.functions.balanceOf(userWrapAndNiftyswap.address)
         const wrappedTokenBalance = await userTokenWrapper.functions.balanceOf(userWrapAndNiftyswap.address, currencyID)
 
-        let addresses = new Array(nTokenTypes).fill('').map((a, i) => userWrapAndNiftyswap.address)
+        const addresses = new Array(nTokenTypes).fill('').map(() => userWrapAndNiftyswap.address)
         const erc1155Balances = await userERC1155Contract.functions.balanceOfBatch(addresses, types)
 
         expect(erc20Balance[0]).to.be.eql(ethers.constants.Zero)
