@@ -1,6 +1,6 @@
 // List all specs here
-require("./NiftswapExchange.spec")
-require("./NiftswapFactory.spec")
-require("./WrapAndNiftyswap.spec")
-require("./NiftswapExchange20.spec")
-require("./NiftswapFactory20.spec")
+require('./NiftswapExchange.spec')
+require('./NiftswapFactory.spec')
+require('./WrapAndNiftyswap.spec')
+require('./NiftswapExchange20.spec')
+require('./NiftswapFactory20.spec')
