@@ -12,7 +12,7 @@ interface NetworkMapping {
 
 interface BuildArtifact {
   readonly contractName: string
-  readonly abi: any[]
+  readonly abi: any[] // eslint-disable-line @typescript-eslint/no-explicit-any
   readonly bytecode: string
   readonly networks?: NetworkMapping
 }
@@ -47,15 +47,9 @@ export class AbstractContract {
   public static async fromBuildArtifact(
     buildArtifact: BuildArtifact,
     links?: { [name: string]: Promise<AbstractContract> },
-    artifactName: string = 'UntitledContract'
+    artifactName = 'UntitledContract'
   ): Promise<AbstractContract> {
-    return new AbstractContract(
-      buildArtifact.abi,
-      buildArtifact.bytecode,
-      buildArtifact.networks,
-      links,
-      artifactName
-    )
+    return new AbstractContract(buildArtifact.abi, buildArtifact.bytecode, buildArtifact.networks, links, artifactName)
   }
 
   public static async getNetworkID(wallet: ethers.Wallet): Promise<number> {
@@ -98,19 +92,14 @@ export class AbstractContract {
    * @returns New contract instance
    */
   public async deploy(wallet: ethers.Wallet, args?: any[]): Promise<ethers.Contract> {
+    // eslint-disable-line @typescript-eslint/no-explicit-any
     if (!wallet.provider) {
       throw new Error('Signer requires provider')
     }
 
     const networkId = (await wallet.provider.getNetwork()).chainId
-    const bytecode = (await this.links)
-      ? await this.generateLinkedBytecode(networkId)
-      : this.bytecode
-    const contractFactory = new ethers.ContractFactory(
-      this.abi,
-      bytecode,
-      wallet
-    )
+    const bytecode = (await this.links) ? await this.generateLinkedBytecode(networkId) : this.bytecode
+    const contractFactory = new ethers.ContractFactory(this.abi, bytecode, wallet)
     return contractFactory.deploy(...(args || []))
   }
 
@@ -120,21 +109,14 @@ export class AbstractContract {
    * @param address Address of deployed instance to connect to
    * @returns Contract instance
    */
-  public async connect(
-    signer: ethers.Signer,
-    address: string
-  ): Promise<ethers.Contract> {
+  public async connect(signer: ethers.Signer, address: string): Promise<ethers.Contract> {
     return new ethers.Contract(address, this.abi, signer)
   }
 
   public getDeployedAddress(networkId: number): string {
     const info = this.networks ? this.networks[networkId] : null
     if (!info) {
-      throw new Error(
-        `Abstract contract ${
-          this.contractName
-        } not deployed on network ${networkId}`
-      )
+      throw new Error(`Abstract contract ${this.contractName} not deployed on network ${networkId}`)
     }
     return info.address
   }
