@@ -1,25 +1,22 @@
-import * as ethers from 'ethers'
+import { ethers, utils, BigNumber } from 'ethers'
 
-export const UNIT_ETH = ethers.utils.parseEther('1')
-export const HIGH_GAS_LIMIT = { gasLimit: 6e9 }
+export const UNIT_ETH = utils.parseEther('1')
+export const HIGH_GAS_LIMIT = { gasLimit: 6_000_000 }
 export const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000'
 
-import { 
-  BuyTokensObj, 
-  SellTokensObj, 
-  AddLiquidityObj, 
-  RemoveLiquidityObj, 
-  SellTokensObj20,
-} from 'src/typings/tx-types'
-import { BigNumber } from 'ethers';
+import { BuyTokensObj, SellTokensObj, AddLiquidityObj, RemoveLiquidityObj, SellTokensObj20 } from 'src/typings/tx-types'
 
 // createTestWallet creates a new wallet
-export const createTestWallet = (web3: any, addressIndex: number = 0) => {
+export const createTestWallet = (web3, addressIndex = 0) => {
   const provider = new Web3DebugProvider(web3.currentProvider)
 
-  const wallet = ethers.Wallet
-    .fromMnemonic(process.env.npm_package_config_mnemonic!, `m/44'/60'/0'/0/${addressIndex}`)
-    .connect(provider)
+  const { npm_package_config_mnemonic } = process.env
+  if (!npm_package_config_mnemonic) {
+    console.error('Missing npm_package_config_mnemonic')
+    process.exit(1)
+  }
+
+  const wallet = ethers.Wallet.fromMnemonic(npm_package_config_mnemonic, `m/44'/60'/0'/0/${addressIndex}`).connect(provider)
 
   const signer = provider.getSigner(addressIndex)
 
@@ -28,24 +25,30 @@ export const createTestWallet = (web3: any, addressIndex: number = 0) => {
 
 // Check if tx was Reverted with specified message
 export function RevertError(errorMessage?: string) {
-  let prefix = 'VM Exception while processing transaction: revert'
-  return errorMessage ? RegExp(`^${prefix + ' ' + errorMessage}$`) : RegExp(`^${prefix}$`)
+  if (!errorMessage) {
+    return /Transaction reverted and Hardhat couldn't infer the reason/
+  } else {
+    // return new RegExp(`${errorMessage}`)
+    return new RegExp(`VM Exception while processing transaction: reverted with reason string ["']${errorMessage}["']`)
+  }
 }
-// In some cases hardhat throws invalid op code when Ganache throws revert
+export const CallError = () => /call revert exception/
 export const OpCodeError = () => RegExp('^VM Exception while processing transaction: (revert|invalid opcode)$')
+export const ArrayAccessError = () => /Array accessed at an out-of-bounds or negative index/
+export const RevertUnsafeMathError = () => /Arithmetic operation .*flowed/
 
 export const methodsSignature = {
-  BUYTOKENS: "0xb2d81047",
-  SELLTOKENS: "0xdb08ec97",
-  ADDLIQUIDITY: "0x82da2b73",
-  REMOVELIQUIDITY: "0x5c0bf259"
+  BUYTOKENS: '0xb2d81047',
+  SELLTOKENS: '0xdb08ec97',
+  ADDLIQUIDITY: '0x82da2b73',
+  REMOVELIQUIDITY: '0x5c0bf259'
 }
 
 export const methodsSignature20 = {
-  BUYTOKENS: "0xb2d81047",
-  SELLTOKENS: "0xade79c7a",
-  ADDLIQUIDITY: "0x82da2b73",
-  REMOVELIQUIDITY: "0x5c0bf259"
+  BUYTOKENS: '0xb2d81047',
+  SELLTOKENS: '0xade79c7a',
+  ADDLIQUIDITY: '0x82da2b73',
+  REMOVELIQUIDITY: '0x5c0bf259'
 }
 
 export const BuyTokensType = `tuple(
@@ -89,9 +92,10 @@ export interface JSONRPCRequest {
 
 export const getBuyTokenData = (
   recipient: string,
-  types: number[] | BigNumber[], 
+  types: number[] | BigNumber[],
   tokensAmountsToBuy: BigNumber[],
-  deadline: number) => {
+  deadline: number
+) => {
   const buyTokenObj = {
     recipient: recipient,
     tokensBoughtIDs: types,
@@ -99,30 +103,23 @@ export const getBuyTokenData = (
     deadline: deadline
   } as BuyTokensObj
 
-
-  return ethers.utils.defaultAbiCoder.encode(
-    ['bytes4', BuyTokensType], [methodsSignature.BUYTOKENS, buyTokenObj])
+  return ethers.utils.defaultAbiCoder.encode(['bytes4', BuyTokensType], [methodsSignature.BUYTOKENS, buyTokenObj])
 }
 
-export const getSellTokenData = (
-  recipient: string,
-  cost: BigNumber, 
-  deadline: number
-) => {
+export const getSellTokenData = (recipient: string, cost: BigNumber, deadline: number) => {
   const sellTokenObj = {
     recipient: recipient,
     minBaseTokens: cost,
     deadline: deadline
   } as SellTokensObj
 
-  return ethers.utils.defaultAbiCoder.encode(
-    ['bytes4', SellTokensType], [methodsSignature.SELLTOKENS, sellTokenObj])
+  return ethers.utils.defaultAbiCoder.encode(['bytes4', SellTokensType], [methodsSignature.SELLTOKENS, sellTokenObj])
 }
 
 // Buy and sell data for ERC-20 exchange
 export const getSellTokenData20 = (
   recipient: string,
-  cost: BigNumber, 
+  cost: BigNumber,
   deadline: number,
   extraFeeRecipients?: string[],
   extraFeeAmounts?: BigNumber[]
@@ -132,29 +129,28 @@ export const getSellTokenData20 = (
     minCurrency: cost,
     extraFeeRecipients: extraFeeRecipients ? extraFeeRecipients : [],
     extraFeeAmounts: extraFeeAmounts ? extraFeeAmounts : [],
-    deadline: deadline,
+    deadline: deadline
   } as SellTokensObj20
 
-  return ethers.utils.defaultAbiCoder.encode(
-    ['bytes4', SellTokens20Type], [methodsSignature20.SELLTOKENS, sellTokenObj])
+  return ethers.utils.defaultAbiCoder.encode(['bytes4', SellTokens20Type], [methodsSignature20.SELLTOKENS, sellTokenObj])
 }
 
 export const getAddLiquidityData = (maxCurrency: BigNumber[], deadline: number) => {
-  const addLiquidityObj = {maxCurrency, deadline} as AddLiquidityObj
+  const addLiquidityObj = { maxCurrency, deadline } as AddLiquidityObj
 
-  return ethers.utils.defaultAbiCoder.encode(
-    ['bytes4', AddLiquidityType], [methodsSignature20.ADDLIQUIDITY, addLiquidityObj])
+  return ethers.utils.defaultAbiCoder.encode(['bytes4', AddLiquidityType], [methodsSignature20.ADDLIQUIDITY, addLiquidityObj])
 }
 
 export const getRemoveLiquidityData = (minCurrency: BigNumber[], minTokens: BigNumber[], deadline: number) => {
-  const removeLiquidityObj = { minCurrency, minTokens, deadline} as RemoveLiquidityObj
+  const removeLiquidityObj = { minCurrency, minTokens, deadline } as RemoveLiquidityObj
 
   return ethers.utils.defaultAbiCoder.encode(
-    ['bytes4', RemoveLiquidityType], [methodsSignature20.REMOVELIQUIDITY, removeLiquidityObj])
+    ['bytes4', RemoveLiquidityType],
+    [methodsSignature20.REMOVELIQUIDITY, removeLiquidityObj]
+  )
 }
 
 export class Web3DebugProvider extends ethers.providers.JsonRpcProvider {
-
   public reqCounter = 0
   public reqLog: JSONRPCRequest[] = []
 
@@ -162,30 +158,29 @@ export class Web3DebugProvider extends ethers.providers.JsonRpcProvider {
   private _sendAsync: (request: any, callback: (error: any, response: any) => void) => void
 
   constructor(web3Provider: ethers.providers.ExternalProvider, network?: ethers.providers.Networkish) {
-      // HTTP has a host; IPC has a path.
-      super(web3Provider.host || web3Provider.path || '', network)
+    // HTTP has a host; IPC has a path.
+    super(web3Provider.host || web3Provider.path || '', network)
 
-      if (web3Provider) {
-        if (web3Provider.sendAsync) {
-          this._sendAsync = web3Provider.sendAsync.bind(web3Provider)
-        } else if (web3Provider.send) {
-          this._sendAsync = web3Provider.send.bind(web3Provider)
-        }
+    if (web3Provider) {
+      if (web3Provider.sendAsync) {
+        this._sendAsync = web3Provider.sendAsync.bind(web3Provider)
+      } else if (web3Provider.send) {
+        this._sendAsync = web3Provider.send.bind(web3Provider)
       }
+    }
 
-      if (!web3Provider || !this._sendAsync) {
-          console.error(ethers.errors.INVALID_ARGUMENT)
-      }
+    if (!web3Provider || !this._sendAsync) {
+      console.error(ethers.errors.INVALID_ARGUMENT)
+    }
 
-      ethers.utils.defineReadOnly(this, '_web3Provider', web3Provider)
+    ethers.utils.defineReadOnly(this, '_web3Provider', web3Provider)
   }
 
   send(method: string, params: any): Promise<any> {
-
     this.reqCounter++
 
     return new Promise((resolve, reject) => {
-      let request = {
+      const request = {
         method: method,
         params: params,
         id: this.reqCounter,
@@ -193,15 +188,14 @@ export class Web3DebugProvider extends ethers.providers.JsonRpcProvider {
       } as JSONRPCRequest
       this.reqLog.push(request)
 
-      this._sendAsync(request, function(error, result) {
+      this._sendAsync(request, function (error, result) {
         if (error) {
           reject(error)
           return
         }
 
         if (result.error) {
-          // @TODO: not any
-          let error: any = new Error(result.error.message)
+          const error = new Error(result.error.message)
           error.code = result.error.code
           error.data = result.error.data
           reject(error)
@@ -213,11 +207,10 @@ export class Web3DebugProvider extends ethers.providers.JsonRpcProvider {
     })
   }
 
-  getPastRequest(reverseIndex: number = 0): JSONRPCRequest {
+  getPastRequest(reverseIndex = 0): JSONRPCRequest {
     if (this.reqLog.length === 0) {
       return { jsonrpc: '2.0', id: 0, method: null, params: null }
     }
-    return this.reqLog[this.reqLog.length-reverseIndex-1]
+    return this.reqLog[this.reqLog.length - reverseIndex - 1]
   }
-
 }
